--- showcase/scripts/__tests__/audit.test.ts
@@ -1,3034 +1,0 @@
-import { describe, it, expect, beforeEach, afterEach, vi } from "vitest";
-import fs from "fs";
-import path from "path";
-import os from "os";
-import { fileURLToPath } from "url";
-import { spawnSync } from "child_process";
-import {
-  auditPackage,
-  buildReport,
-  computeExitCode,
-  listShowcasePackageSlugs,
-  readManifest,
-  countFiles,
-  findExamplesSource,
-  resolveExamplesSource,
-  isProgrammerBug,
-  parseArgs,
-  anomalyMessage,
-  UnreadableDirError,
-  canonicalizeForIsMain,
-  BORN_IN_SHOWCASE,
-  SLUG_TO_EXAMPLES,
-  type AuditConfig,
-  type Anomaly,
-  type PackageAudit,
-} from "../audit.js";
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-
-const AUDIT_SCRIPT = path.resolve(__dirname, "..", "audit.ts");
-
-/**
- * Build a throwaway temp tree mimicking:
- *   <root>/packages/<slug>/manifest.yaml
- *   <root>/packages/<slug>/tests/e2e/*.spec.ts
- *   <root>/packages/<slug>/qa/*.md
- *   <root>/examples/integrations/<name>/
- */
-function makeTmpTree(): string {
-  const root = fs.mkdtempSync(path.join(os.tmpdir(), "audit-fixture-"));
-  fs.mkdirSync(path.join(root, "packages"), { recursive: true });
-  fs.mkdirSync(path.join(root, "examples", "integrations"), {
-    recursive: true,
-  });
-  return root;
-}
-
-function makeConfig(root: string): AuditConfig {
-  return {
-    packagesDir: path.join(root, "packages"),
-    examplesIntegrationsDir: path.join(root, "examples", "integrations"),
-    repoRoot: root,
-  };
-}
-
-function writePackage(
-  root: string,
-  slug: string,
-  opts: {
-    manifest?: string; // raw YAML string; undefined = no manifest.yaml
-    specs?: string[];
-    qaFiles?: string[];
-  },
-) {
-  const pkgDir = path.join(root, "packages", slug);
-  fs.mkdirSync(pkgDir, { recursive: true });
-  if (opts.manifest !== undefined) {
-    fs.writeFileSync(path.join(pkgDir, "manifest.yaml"), opts.manifest);
-  }
-  if (opts.specs && opts.specs.length > 0) {
-    const e2eDir = path.join(pkgDir, "tests", "e2e");
-    fs.mkdirSync(e2eDir, { recursive: true });
-    for (const s of opts.specs) {
-      fs.writeFileSync(path.join(e2eDir, s), "// test\n");
-    }
-  }
-  if (opts.qaFiles && opts.qaFiles.length > 0) {
-    const qaDir = path.join(pkgDir, "qa");
-    fs.mkdirSync(qaDir, { recursive: true });
-    for (const q of opts.qaFiles) {
-      fs.writeFileSync(path.join(qaDir, q), "# qa\n");
-    }
-  }
-}
-
-function makeExampleDir(root: string, name: string) {
-  fs.mkdirSync(path.join(root, "examples", "integrations", name), {
-    recursive: true,
-  });
-}
-
-// Helpers that recover the old string-based predicates so tests read like
-// a behavioral spec even though the underlying type is now tagged.
-function anomalyStrings(a: PackageAudit): string[] {
-  return a.anomalies.map(anomalyMessage);
-}
-
-describe("readManifest", () => {
-  let root: string;
-  beforeEach(() => {
-    root = makeTmpTree();
-  });
-  afterEach(() => {
-    fs.rmSync(root, { recursive: true, force: true });
-  });
-
-  it("returns { kind: 'missing' } when manifest.yaml does not exist", () => {
-    writePackage(root, "mypkg", {});
-    const cfg = makeConfig(root);
-    const r = readManifest("mypkg", cfg);
-    expect(r.kind).toBe("missing");
-  });
-
-  it("returns { kind: 'malformed', error } when manifest.yaml is malformed YAML", () => {
-    writePackage(root, "mypkg", {
-      manifest: "slug: mypkg\n  bad:: indent: [unterminated\n",
-    });
-    const cfg = makeConfig(root);
-    const r = readManifest("mypkg", cfg);
-    expect(r.kind).toBe("malformed");
-    if (r.kind === "malformed") {
-      expect(typeof r.error).toBe("string");
-      expect(r.error.length).toBeGreaterThan(0);
-    }
-  });
-
-  it("returns { kind: 'ok', manifest } for a valid manifest", () => {
-    writePackage(root, "mypkg", {
-      manifest: "slug: mypkg\ndeployed: true\ndemos:\n  - id: foo\n",
-    });
-    const cfg = makeConfig(root);
-    const r = readManifest("mypkg", cfg);
-    expect(r.kind).toBe("ok");
-    if (r.kind === "ok") {
-      expect(r.manifest.slug).toBe("mypkg");
-      expect(r.manifest.deployed).toBe(true);
-      expect(r.manifest.demos?.length).toBe(1);
-    }
-  });
-
-  it("returns { kind: 'malformed' } for empty manifest.yaml (yaml.parse → null)", () => {
-    // yaml.parse("") returns null, and previously the parsed-as-Manifest
-    // cast let null propagate into downstream .demos / .deployed access,
-    // crashing with TypeError. Empty/non-object YAML must be rejected here
-    // before auditPackage touches it.
-    writePackage(root, "mypkg", { manifest: "" });
-    const cfg = makeConfig(root);
-    const r = readManifest("mypkg", cfg);
-    expect(r.kind).toBe("malformed");
-  });
-
-  it("returns { kind: 'malformed' } when manifest.yaml parses to a non-object (e.g. bare scalar)", () => {
-    // yaml.parse("42") returns the number 42 — also not a valid Manifest.
-    writePackage(root, "mypkg", { manifest: "42\n" });
-    const cfg = makeConfig(root);
-    const r = readManifest("mypkg", cfg);
-    expect(r.kind).toBe("malformed");
-  });
-
-  it("passes the dir slug to parseManifest so a slug-mismatch flags as malformed", () => {
-    // readManifest must pass the directory slug to parseManifest so
-    // a manifest whose declared `slug:` disagrees
-    // with the directory holding it is surfaced as a shape error
-    // (not silently accepted). Package dir name is "dir-slug" but
-    // manifest declares "manifest-slug" — shape validation catches
-    // the drift only if dirSlug flows through.
-    writePackage(root, "dir-slug", {
-      manifest: "slug: manifest-slug\ndeployed: true\ndemos:\n  - id: a\n",
-    });
-    const cfg = makeConfig(root);
-    const r = readManifest("dir-slug", cfg);
-    expect(r.kind).toBe("malformed");
-    if (r.kind === "malformed") {
-      expect(r.subkind).toBe("shape");
-      expect(r.error).toMatch(/slug mismatch/);
-      expect(r.error).toContain("manifest-slug");
-      expect(r.error).toContain("dir-slug");
-    }
-  });
-
-  it("accepts a manifest whose slug matches the directory name", () => {
-    // Counterpart of the slug-mismatch test: when the manifest's
-    // declared slug matches the directory slug, parseManifest's guard
-    // passes and the kind is 'ok'.
-    writePackage(root, "same-slug", {
-      manifest: "slug: same-slug\ndeployed: true\ndemos:\n  - id: a\n",
-    });
-    const cfg = makeConfig(root);
-    const r = readManifest("same-slug", cfg);
-    expect(r.kind).toBe("ok");
-  });
-
-  it("returns { kind: 'unreadable', error } (distinct from 'malformed') on EACCES", () => {
-    // Contract: audit.ts does not collapse unreadable into malformed
-    // with a string prefix. Downstream switches on all 4 ParsedManifest
-    // variants and classifies unreadable manifests under the
-    // `unreadable` bucket (alongside spec/qa-dir I/O failures), not
-    // under `malformedManifest` (which is content-shape-only).
-    writePackage(root, "mypkg", {
-      manifest: "slug: mypkg\ndeployed: true\ndemos:\n  - id: foo\n",
-    });
-    const target = path.join(root, "packages", "mypkg", "manifest.yaml");
-    const orig = fs.readFileSync;
-    const spy = vi.spyOn(fs, "readFileSync").mockImplementation(((
-      p: fs.PathOrFileDescriptor,
-      options?: unknown,
-    ) => {
-      if (typeof p === "string" && p === target) {
-        const e: NodeJS.ErrnoException = new Error("EACCES: permission denied");
-        e.code = "EACCES";
-        throw e;
-      }
-      return (
-        orig as unknown as (p: fs.PathOrFileDescriptor, o?: unknown) => unknown
-      )(p, options);
-    }) as typeof fs.readFileSync);
-    try {
-      const cfg = makeConfig(root);
-      const r = readManifest("mypkg", cfg);
-      expect(r.kind).toBe("unreadable");
-      if (r.kind === "unreadable") {
-        expect(r.error).toContain("EACCES");
-      }
-      // And buildReport routes it under the `unreadable` bucket, NOT
-      // `malformedManifest`.
-      const report = buildReport(["mypkg"], cfg);
-      expect(report.anomalies.unreadable).toContain("mypkg");
-      expect(report.anomalies.malformedManifest).not.toContain("mypkg");
-    } finally {
-      spy.mockRestore();
-    }
-  });
-});
-
-describe("countFiles", () => {
-  let root: string;
-  beforeEach(() => {
-    root = makeTmpTree();
-  });
-  afterEach(() => {
-    fs.rmSync(root, { recursive: true, force: true });
-  });
-
-  it("returns { state: 'missing' } when dir does not exist", () => {
-    const r = countFiles(path.join(root, "does-not-exist"), (n) =>
-      n.endsWith(".spec.ts"),
-    );
-    expect(r.state).toBe("missing");
-  });
-
-  it("returns { state: 'ok', count: N } for readable dir", () => {
-    const d = path.join(root, "some");
-    fs.mkdirSync(d);
-    fs.writeFileSync(path.join(d, "a.spec.ts"), "");
-    fs.writeFileSync(path.join(d, "b.spec.ts"), "");
-    fs.writeFileSync(path.join(d, "c.md"), "");
-    const r = countFiles(d, (n) => n.endsWith(".spec.ts"));
-    expect(r.state).toBe("ok");
-    if (r.state === "ok") {
-      expect(r.count).toBe(2);
-    }
-  });
-
-  it("surfaces { state: 'unreadable' } instead of silent 0 on readdirSync failure", () => {
-    const d = path.join(root, "unreadable");
-    fs.mkdirSync(d);
-    // Mock readdirSync to throw EACCES for this specific path.
-    const orig = fs.readdirSync;
-    const spy = vi.spyOn(fs, "readdirSync").mockImplementation(((
-      p: fs.PathLike,
-      options?: unknown,
-    ) => {
-      if (typeof p === "string" && p === d) {
-        const e: NodeJS.ErrnoException = new Error("EACCES: permission denied");
-        e.code = "EACCES";
-        throw e;
-      }
-      return (orig as unknown as (p: fs.PathLike, o?: unknown) => unknown)(
-        p,
-        options,
-      );
-    }) as typeof fs.readdirSync);
-    try {
-      const r = countFiles(d, (n) => n.endsWith(".spec.ts"));
-      expect(r.state).toBe("unreadable");
-      if (r.state === "unreadable") {
-        expect(r.error).toContain("EACCES");
-      }
-    } finally {
-      spy.mockRestore();
-    }
-  });
-
-  it("surfaces { state: 'unreadable' } (not 'missing') when statSync throws EACCES", () => {
-    // Regression guard: the old implementation gated on `fs.existsSync`,
-    // which returns false for EACCES/EPERM/EIO/ELOOP/ENOTDIR just like it
-    // does for ENOENT. That silently classified unreadable dirs as
-    // `missing` (legitimate zero) and triggered phantom count-mismatch
-    // anomalies. The fix replaces existsSync with a statSync + errno
-    // branch so non-ENOENT stat failures surface as `unreadable`.
-    const d = path.join(root, "eacces-stat");
-    fs.mkdirSync(d);
-    const orig = fs.statSync;
-    const spy = vi.spyOn(fs, "statSync").mockImplementation(((
-      p: fs.PathLike,
-      options?: unknown,
-    ) => {
-      if (typeof p === "string" && p === d) {
-        const e: NodeJS.ErrnoException = new Error(
-          "EACCES: permission denied (statSync)",
-        );
-        e.code = "EACCES";
-        throw e;
-      }
-      return (orig as unknown as (p: fs.PathLike, o?: unknown) => unknown)(
-        p,
-        options,
-      );
-    }) as typeof fs.statSync);
-    try {
-      const r = countFiles(d, (n) => n.endsWith(".spec.ts"));
-      // MUST NOT collapse to `missing` — operator needs the unreadable
-      // signal so downstream auditPackage emits `unreadable-dir` instead
-      // of a phantom count-mismatch on a zero count.
-      expect(r.state).toBe("unreadable");
-      if (r.state === "unreadable") {
-        expect(r.error).toContain("EACCES");
-      }
-    } finally {
-      spy.mockRestore();
-    }
-  });
-});
-
-describe("auditPackage — EACCES on spec dir stat → unreadable-dir anomaly (not count-mismatch)", () => {
-  let root: string;
-  beforeEach(() => {
-    root = makeTmpTree();
-  });
-  afterEach(() => {
-    fs.rmSync(root, { recursive: true, force: true });
-  });
-
-  it("EACCES on statSync of tests/e2e → unreadable-dir anomaly, no phantom count-mismatch", () => {
-    // End-to-end: when the e2e dir's stat fails with EACCES, the package
-    // must surface an `unreadable-dir` anomaly — NOT a `count-mismatch`
-    // claiming 0 specs against declared demos. Previously the existsSync
-    // gate in countFiles collapsed EACCES to `missing` (count=0), which
-    // then compared unequal against demos and fired a phantom mismatch.
-    writePackage(root, "eacces-e2e", {
-      manifest: `slug: eacces-e2e\ndeployed: true\ndemos:\n  - id: a\n`,
-      specs: ["a.spec.ts"],
-      qaFiles: ["a.md"],
-    });
-    const e2eDir = path.join(root, "packages", "eacces-e2e", "tests", "e2e");
-    const orig = fs.statSync;
-    const spy = vi.spyOn(fs, "statSync").mockImplementation(((
-      p: fs.PathLike,
-      options?: unknown,
-    ) => {
-      if (typeof p === "string" && p === e2eDir) {
-        const e: NodeJS.ErrnoException = new Error("EACCES: injected");
-        e.code = "EACCES";
-        throw e;
-      }
-      return (orig as unknown as (p: fs.PathLike, o?: unknown) => unknown)(
-        p,
-        options,
-      );
-    }) as typeof fs.statSync);
-    try {
-      const cfg = makeConfig(root);
-      const a = auditPackage("eacces-e2e", cfg);
-      // unreadable-dir anomaly MUST surface with an error containing EACCES.
-      const unreadable = a.anomalies.find(
-        (x): x is Extract<Anomaly, { kind: "unreadable-dir" }> =>
-          x.kind === "unreadable-dir",
-      );
-      expect(
-        unreadable,
-        `expected unreadable-dir anomaly, got: ${JSON.stringify(a.anomalies)}`,
-      ).toBeDefined();
-      if (unreadable) {
-        expect(unreadable.error).toContain("EACCES");
-      }
-      // MUST NOT fire a phantom spec count-mismatch (unreadable !== zero).
-      const specMismatch = a.anomalies.find(
-        (x): x is Extract<Anomaly, { kind: "count-mismatch" }> =>
-          x.kind === "count-mismatch" && x.dimension === "spec",
-      );
-      expect(specMismatch).toBeUndefined();
-      // spec CountState must be "unreadable" (not "missing").
-      expect(a.spec.state).toBe("unreadable");
-    } finally {
-      spy.mockRestore();
-    }
-  });
-});
-
-describe("findExamplesSource", () => {
-  let root: string;
-  beforeEach(() => {
-    root = makeTmpTree();
-  });
-  afterEach(() => {
-    fs.rmSync(root, { recursive: true, force: true });
-  });
-
-  it("returns null source when no candidate directory exists", () => {
-    const cfg = makeConfig(root);
-    const r = findExamplesSource("does-not-exist", cfg);
-    expect(r.source).toBeNull();
-    expect(r.unreadableForSlug).toBe(false);
-  });
-
-  it("returns relative path when a candidate dir exists", () => {
-    makeExampleDir(root, "crewai-crews");
-    const cfg = makeConfig(root);
-    const r = findExamplesSource("crewai-crews", cfg);
-    expect(r.source).toBe(
-      path.join("examples", "integrations", "crewai-crews"),
-    );
-    expect(r.unreadableForSlug).toBe(false);
-  });
-
-  it("does not crash if statSync throws — treats error as not-found", () => {
-    makeExampleDir(root, "crewai-crews");
-    const target = path.join(root, "examples", "integrations", "crewai-crews");
-    const orig = fs.statSync;
-    const spy = vi.spyOn(fs, "statSync").mockImplementation(((
-      p: fs.PathLike,
-      options?: unknown,
-    ) => {
-      if (typeof p === "string" && p === target) {
-        const e: NodeJS.ErrnoException = new Error("EIO: race condition");
-        e.code = "EIO";
-        throw e;
-      }
-      return (orig as unknown as (p: fs.PathLike, o?: unknown) => unknown)(
-        p,
-        options,
-      );
-    }) as typeof fs.statSync);
-    try {
-      const cfg = makeConfig(root);
-      const r = findExamplesSource("crewai-crews", cfg);
-      expect(r.source).toBeNull();
-    } finally {
-      spy.mockRestore();
-    }
-  });
-
-  it("records stale-mapping warnings on the supplied sink array (no stderr monkey-patch)", () => {
-    // Regression guard for the old global-state approach: findExamplesSource
-    // used to monkey-patch process.stderr.write and restore it in a
-    // finally block. Concurrent calls would collide, and any caller that
-    // captured stderr (e.g. vitest's stderr spy) saw nothing because
-    // the monkey-patch sandwich restored the original. The new contract:
-    // pass an explicit string[] sink; warnings are appended to it, and
-    // the function never touches process.stderr.
-    const mappedSlug = "mastra";
-    expect(SLUG_TO_EXAMPLES[mappedSlug]).toBeDefined();
-    const cfg = makeConfig(root);
-    const sink: string[] = [];
-    const r = findExamplesSource(mappedSlug, cfg, sink);
-    expect(r.source).toBeNull();
-    expect(r.unreadableForSlug).toBe(false);
-    // Sink received the stale-mapping warning verbatim.
-    expect(sink.length).toBeGreaterThan(0);
-    expect(sink.some((w) => w.includes(mappedSlug))).toBe(true);
-    expect(sink.some((w) => /warn/i.test(w))).toBe(true);
-  });
-
-  it("does not push to the sink when the slug is unmapped (fallback path)", () => {
-    // Unmapped slug falls back to [slug] — that's the "no mapping
-    // needed" path, not a dead entry, so no warning.
-    const cfg = makeConfig(root);
-    const sink: string[] = [];
-    const r = findExamplesSource("totally-unmapped-slug", cfg, sink);
-    expect(r.source).toBeNull();
-    expect(sink).toEqual([]);
-  });
-
-  it("statSync failures land on the sink (not stderr)", () => {
-    makeExampleDir(root, "crewai-crews");
-    const target = path.join(root, "examples", "integrations", "crewai-crews");
-    const orig = fs.statSync;
-    const spy = vi.spyOn(fs, "statSync").mockImplementation(((
-      p: fs.PathLike,
-      options?: unknown,
-    ) => {
-      if (typeof p === "string" && p === target) {
-        const e: NodeJS.ErrnoException = new Error("EIO: injected");
-        e.code = "EIO";
-        throw e;
-      }
-      return (orig as unknown as (p: fs.PathLike, o?: unknown) => unknown)(
-        p,
-        options,
-      );
-    }) as typeof fs.statSync);
-    try {
-      const cfg = makeConfig(root);
-      const sink: string[] = [];
-      const r = findExamplesSource("crewai-crews", cfg, sink);
-      expect(r.source).toBeNull();
-      expect(sink.some((w) => w.includes("statSync"))).toBe(true);
-      expect(sink.some((w) => w.includes("EIO"))).toBe(true);
-    } finally {
-      spy.mockRestore();
-    }
-  });
-
-  it("returns the first declared candidate when only the first exists", () => {
-    const slug = "langgraph-typescript";
-    const mapped = SLUG_TO_EXAMPLES[slug];
-    expect(mapped).toBeDefined();
-    expect(mapped!.length).toBeGreaterThan(0);
-    const first = mapped![0];
-    makeExampleDir(root, first);
-    const cfg = makeConfig(root);
-    const r = findExamplesSource(slug, cfg);
-    expect(r.source).toBe(path.join("examples", "integrations", first));
-    expect(r.unreadableForSlug).toBe(false);
-  });
-
-  it("falls back to a later candidate when only that later candidate exists", () => {
-    // Exercise the multi-candidate fallback path via resolveExamplesSource,
-    // which accepts an explicit `mapped` tuple. All live SLUG_TO_EXAMPLES
-    // entries are single-candidate today, so using findExamplesSource
-    // would short-circuit the fallback loop and leave this codepath
-    // uncovered. Synthetic multi-candidate input makes the assertion
-    // meaningful regardless of the production map's shape.
-    const slug = "synthetic-multi";
-    const mapped = ["missing-first", "real-later"] as const;
-    makeExampleDir(root, "real-later");
-    const cfg = makeConfig(root);
-    const r = resolveExamplesSource(slug, mapped, cfg);
-    expect(r.source).toBe(path.join("examples", "integrations", "real-later"));
-    expect(r.unreadableForSlug).toBe(false);
-  });
-
-  it("warns and returns null when an unmapped slug's candidate path exists but is a regular file", () => {
-    // An unmapped slug whose candidate path resolves to a regular file
-    // (stray file in examples/integrations, or a name collision) must
-    // warn rather than return null silently — operators need a signal
-    // that a seemingly-present path was skipped. The warning wording
-    // ("exists but is not a directory") is distinct from the
-    // mapped-entry "no matching directory" warning so the two
-    // misconfiguration modes stay disambiguable.
-    const slug = "unmapped-file-slug";
-    // Write a regular file at examples/integrations/<slug> (no mkdir).
-    const full = path.join(root, "examples", "integrations", slug);
-    fs.writeFileSync(full, "not a directory\n");
-    const cfg = makeConfig(root);
-    const sink: string[] = [];
-    // mapped = undefined exercises the unmapped-slug branch.
-    const r = resolveExamplesSource(slug, undefined, cfg, sink);
-    expect(r.source).toBeNull();
-    expect(r.unreadableForSlug).toBe(false);
-    expect(
-      sink.some(
-        (w) => w.includes("exists but is not a directory") && w.includes(full),
-      ),
-      `expected file-not-dir warning for ${full} in sink: ${JSON.stringify(
-        sink,
-      )}`,
-    ).toBe(true);
-    // Must NOT fire the mapped-entry "no matching directory" warning —
-    // the fallback path is explicitly not a mapping.
-    expect(sink.some((w) => w.includes("SLUG_TO_EXAMPLES entry"))).toBe(false);
-  });
-
-  it("EACCES on existsSync does NOT silently skip mapped candidate (routes to unreadableForSlug)", () => {
-    // Regression guard: the old implementation used `fs.existsSync` to
-    // pre-filter candidates. existsSync returns false for EACCES just
-    // like it does for ENOENT, so a candidate whose parent dir is
-    // unreadable was silently skipped — not counted in `existedCount`,
-    // not counted in `unreadableCount`. When ALL mapped candidates were
-    // EACCES'd, the resolver returned `unreadableForSlug: false` and the
-    // package silently fell through to `missing-examples`.
-    //
-    // Fix: replace existsSync with a statSync inside try/catch. ENOENT →
-    // continue (absent); EACCES/other errno → increment unreadableCount
-    // and push diagnostic. Result: unreadableForSlug is truthy when every
-    // mapped candidate is EACCES'd.
-    const slug = "synthetic-eacces";
-    const mapped = ["only-cand"] as const;
-    // Do NOT create the candidate path on disk. Mock existsSync so the
-    // old code path would see false (as if ENOENT), while statSync is
-    // configured to throw EACCES for the candidate path — which is the
-    // signal the fixed code must use.
-    const full = path.join(root, "examples", "integrations", "only-cand");
-    const origExists = fs.existsSync;
-    const existsSpy = vi.spyOn(fs, "existsSync").mockImplementation(((
-      p: fs.PathLike,
-    ) => {
-      if (typeof p === "string" && p === full) {
-        // Simulate existsSync hiding EACCES as false — the very bug we
-        // are guarding against.
-        return false;
-      }
-      return (origExists as (p: fs.PathLike) => boolean)(p);
-    }) as typeof fs.existsSync);
-    const origStat = fs.statSync;
-    const statSpy = vi.spyOn(fs, "statSync").mockImplementation(((
-      p: fs.PathLike,
-      options?: unknown,
-    ) => {
-      if (typeof p === "string" && p === full) {
-        const e: NodeJS.ErrnoException = new Error("EACCES: parent unreadable");
-        e.code = "EACCES";
-        throw e;
-      }
-      return (origStat as unknown as (p: fs.PathLike, o?: unknown) => unknown)(
-        p,
-        options,
-      );
-    }) as typeof fs.statSync);
-    try {
-      const cfg = makeConfig(root);
-      const sink: string[] = [];
-      const r = resolveExamplesSource(slug, mapped, cfg, sink);
-      // MUST classify as unreadableForSlug — the operator needs the
-      // infrastructure-failure signal instead of the benign
-      // missing-examples classification.
-      expect(r.source).toBeNull();
-      expect(r.unreadableForSlug).toBe(true);
-      // A diagnostic must land on the sink so the EACCES is not silent.
-      expect(sink.some((w) => w.includes("EACCES"))).toBe(true);
-    } finally {
-      existsSpy.mockRestore();
-      statSpy.mockRestore();
-    }
-  });
-
-  it("returns the first candidate when both first and later candidates exist (first wins)", () => {
-    // Declared order decides — the first present candidate wins
-    // regardless of filesystem enumeration order. Uses synthetic
-    // multi-candidate mapping for the same reason as the fallback test
-    // above: the live map is single-candidate everywhere.
-    const slug = "synthetic-multi";
-    const mapped = ["first-dir", "later-dir"] as const;
-    makeExampleDir(root, "first-dir");
-    makeExampleDir(root, "later-dir");
-    const cfg = makeConfig(root);
-    const r = resolveExamplesSource(slug, mapped, cfg);
-    expect(r.source).toBe(path.join("examples", "integrations", "first-dir"));
-    expect(r.unreadableForSlug).toBe(false);
-  });
-});
-
-describe("auditPackage", () => {
-  let root: string;
-  beforeEach(() => {
-    root = makeTmpTree();
-  });
-  afterEach(() => {
-    fs.rmSync(root, { recursive: true, force: true });
-  });
-
-  it("emits 'malformed manifest.yaml' (not 'missing') for bad YAML", () => {
-    writePackage(root, "broken", {
-      manifest: "demos: [[[\nunterminated\n",
-    });
-    const cfg = makeConfig(root);
-    const a = auditPackage("broken", cfg);
-    expect(a.manifest.kind).toBe("malformed");
-    // The full variant is preserved — callers can reach the underlying
-    // error without a second lookup.
-    if (a.manifest.kind === "malformed") {
-      expect(a.manifest.error).toMatch(/unterminated|Flow sequence|\[/);
-    }
-    expect(a.anomalies.some((x) => x.kind === "malformed-manifest")).toBe(true);
-    expect(a.anomalies.some((x) => x.kind === "missing-manifest")).toBe(false);
-  });
-
-  it("emits 'missing manifest.yaml' when no manifest.yaml exists", () => {
-    writePackage(root, "noman", {});
-    const cfg = makeConfig(root);
-    const a = auditPackage("noman", cfg);
-    expect(a.manifest.kind).toBe("missing");
-    expect(anomalyStrings(a)).toContain("missing manifest.yaml");
-  });
-
-  it("does not crash and emits 'malformed manifest.yaml' for an empty manifest.yaml", () => {
-    // yaml.parse("") → null; if the guard is missing, auditPackage will
-    // throw TypeError when it reads manifest.demos / manifest.deployed.
-    writePackage(root, "empty", { manifest: "" });
-    const cfg = makeConfig(root);
-    expect(() => auditPackage("empty", cfg)).not.toThrow();
-    const a = auditPackage("empty", cfg);
-    expect(a.manifest.kind).toBe("malformed");
-    expect(a.anomalies.some((x) => x.kind === "malformed-manifest")).toBe(true);
-  });
-
-  it("does not flag missingExamples for born-in-showcase slugs", () => {
-    const slug = "ag2"; // known born-in-showcase
-    expect(BORN_IN_SHOWCASE.has(slug)).toBe(true);
-
-    writePackage(root, slug, {
-      manifest: `slug: ${slug}\ndeployed: true\ndemos:\n  - id: x\n`,
-      specs: ["x.spec.ts"],
-      qaFiles: ["x.md"],
-    });
-    // Intentionally DO NOT create an examples/integrations/ag2 directory.
-    const cfg = makeConfig(root);
-    const a = auditPackage(slug, cfg);
-    expect(a.anomalies.some((x) => x.kind === "missing-examples")).toBe(false);
-  });
-
-  it("born-in-showcase + deployed:true → zero anomalies (clean)", () => {
-    const slug = "claude-sdk-typescript";
-    expect(BORN_IN_SHOWCASE.has(slug)).toBe(true);
-    writePackage(root, slug, {
-      manifest: `slug: ${slug}\ndeployed: true\ndemos:\n  - id: x\n`,
-      specs: ["x.spec.ts"],
-      qaFiles: ["x.md"],
-    });
-    const cfg = makeConfig(root);
-    const a = auditPackage(slug, cfg);
-    expect(a.anomalies).toEqual([]);
-  });
-
-  it("surfaces 'could not read' anomaly when spec dir readdir fails", () => {
-    writePackage(root, "perm", {
-      manifest: `slug: perm\ndeployed: true\ndemos:\n  - id: x\n`,
-      specs: ["x.spec.ts"],
-      qaFiles: ["x.md"],
-    });
-    const e2eDir = path.join(root, "packages", "perm", "tests", "e2e");
-    const orig = fs.readdirSync;
-    const spy = vi.spyOn(fs, "readdirSync").mockImplementation(((
-      p: fs.PathLike,
-      options?: unknown,
-    ) => {
-      if (typeof p === "string" && p === e2eDir) {
-        const e: NodeJS.ErrnoException = new Error("EACCES");
-        e.code = "EACCES";
-        throw e;
-      }
-      return (orig as unknown as (p: fs.PathLike, o?: unknown) => unknown)(
-        p,
-        options,
-      );
-    }) as typeof fs.readdirSync);
-    try {
-      const cfg = makeConfig(root);
-      const a = auditPackage("perm", cfg);
-      expect(a.anomalies.some((x) => x.kind === "unreadable-dir")).toBe(true);
-    } finally {
-      spy.mockRestore();
-    }
-  });
-
-  it("surfaces 'could not read' anomaly when qa dir readdir fails (symmetric to spec)", () => {
-    writePackage(root, "qaperm", {
-      manifest: `slug: qaperm\ndeployed: true\ndemos:\n  - id: x\n`,
-      specs: ["x.spec.ts"],
-      qaFiles: ["x.md"],
-    });
-    const qaDir = path.join(root, "packages", "qaperm", "qa");
-    const orig = fs.readdirSync;
-    const spy = vi.spyOn(fs, "readdirSync").mockImplementation(((
-      p: fs.PathLike,
-      options?: unknown,
-    ) => {
-      if (typeof p === "string" && p === qaDir) {
-        const e: NodeJS.ErrnoException = new Error("EACCES");
-        e.code = "EACCES";
-        throw e;
-      }
-      return (orig as unknown as (p: fs.PathLike, o?: unknown) => unknown)(
-        p,
-        options,
-      );
-    }) as typeof fs.readdirSync);
-    try {
-      const cfg = makeConfig(root);
-      const a = auditPackage("qaperm", cfg);
-      const unreadable = a.anomalies.find(
-        (x): x is Extract<Anomaly, { kind: "unreadable-dir" }> =>
-          x.kind === "unreadable-dir",
-      );
-      expect(unreadable).toBeDefined();
-      expect(unreadable!.dir).toBe(qaDir);
-    } finally {
-      spy.mockRestore();
-    }
-  });
-
-  it("deployed:undefined renders as 'deployed=unset' and counts toward notDeployed", () => {
-    // Regression guard: a manifest with no `deployed` field MUST produce
-    // a distinct "unset" state (not collapsed into "explicit-false").
-    // The Anomaly.not-deployed.state carries a self-documenting string
-    // union — callers read the raw boolean off the manifest variant.
-    writePackage(root, "unset", {
-      manifest: `slug: unset\ndemos:\n  - id: x\n`,
-      specs: ["x.spec.ts"],
-      qaFiles: ["x.md"],
-    });
-    makeExampleDir(root, "unset");
-    const cfg = makeConfig(root);
-    const a = auditPackage("unset", cfg);
-    const notDeployed = a.anomalies.find(
-      (x): x is Extract<Anomaly, { kind: "not-deployed" }> =>
-        x.kind === "not-deployed",
-    );
-    expect(notDeployed).toBeDefined();
-    expect(notDeployed!.state).toBe("unset");
-    // Read deployed via the manifest variant — single source of truth.
-    expect(a.manifest.kind).toBe("ok");
-    if (a.manifest.kind === "ok") {
-      expect(a.manifest.manifest.deployed).toBeUndefined();
-    }
-
-    // Counterpart: deployed:false is distinct.
-    writePackage(root, "falsedep", {
-      manifest: `slug: falsedep\ndeployed: false\ndemos:\n  - id: x\n`,
-      specs: ["x.spec.ts"],
-      qaFiles: ["x.md"],
-    });
-    makeExampleDir(root, "falsedep");
-    const b = auditPackage("falsedep", cfg);
-    const bNot = b.anomalies.find(
-      (x): x is Extract<Anomaly, { kind: "not-deployed" }> =>
-        x.kind === "not-deployed",
-    );
-    expect(bNot).toBeDefined();
-    expect(bNot!.state).toBe("explicit-false");
-    expect(b.manifest.kind).toBe("ok");
-    if (b.manifest.kind === "ok") {
-      expect(b.manifest.manifest.deployed).toBe(false);
-    }
-
-    // Both flow into the notDeployed bucket.
-    const report = buildReport(["unset", "falsedep"], cfg);
-    expect([...report.anomalies.notDeployed].sort()).toEqual(
-      ["falsedep", "unset"].sort(),
-    );
-  });
-
-  it("records warnings on the `warnings` field for stale SLUG_TO_EXAMPLES entries", () => {
-    // Contract: even if stderr is captured/redirected (or if the
-    // caller renders JSON and swallows stderr entirely), the audit
-    // record itself should carry the warning.
-    const mappedSlug = "mastra";
-    expect(SLUG_TO_EXAMPLES[mappedSlug]).toBeDefined();
-    writePackage(root, mappedSlug, {
-      manifest: `slug: ${mappedSlug}\ndeployed: true\ndemos:\n  - id: x\n`,
-      specs: ["x.spec.ts"],
-      qaFiles: ["x.md"],
-    });
-    // Intentionally DO NOT create the examples/integrations/<mapped> dir.
-    const cfg = makeConfig(root);
-    const a = auditPackage(mappedSlug, cfg);
-    expect(a.warnings.length).toBeGreaterThan(0);
-    expect(a.warnings.some((w) => w.includes(mappedSlug))).toBe(true);
-  });
-
-  it("pushes an unreadable-examples anomaly (not missing-examples) when all mapped candidates exist but are unreadable", () => {
-    // Contract: the "all candidates unreadable" infrastructure failure
-    // must surface as its own Anomaly variant so downstream consumers
-    // can distinguish "provenance is genuinely missing" (missing-examples,
-    // stale mapping) from "we couldn't tell whether provenance is
-    // satisfied" (unreadable-examples, I/O / permissions). Conflating
-    // them would hide real access failures behind a content-shaped signal.
-    const mappedSlug = "mastra";
-    const candidates = SLUG_TO_EXAMPLES[mappedSlug];
-    expect(candidates).toBeDefined();
-    expect(candidates.length).toBeGreaterThan(0);
-    writePackage(root, mappedSlug, {
-      manifest: `slug: ${mappedSlug}\ndeployed: true\ndemos:\n  - id: x\n`,
-      specs: ["x.spec.ts"],
-      qaFiles: ["x.md"],
-    });
-    // Create the candidate dirs so existsSync sees them; mock statSync
-    // to throw EACCES on those paths so every candidate registers as
-    // unreadable. Non-candidate statSync calls fall through to the real
-    // implementation so other fixture operations still work.
-    const fullPaths = candidates.map((c) =>
-      path.join(root, "examples", "integrations", c),
-    );
-    for (const c of candidates) makeExampleDir(root, c);
-    const orig = fs.statSync;
-    const spy = vi.spyOn(fs, "statSync").mockImplementation(((
-      p: fs.PathLike,
-      options?: unknown,
-    ) => {
-      if (typeof p === "string" && fullPaths.includes(p)) {
-        const e: NodeJS.ErrnoException = new Error("EACCES: injected");
-        e.code = "EACCES";
-        throw e;
-      }
-      return (orig as unknown as (p: fs.PathLike, o?: unknown) => unknown)(
-        p,
-        options,
-      );
-    }) as typeof fs.statSync);
-    try {
-      const cfg = makeConfig(root);
-      const a = auditPackage(mappedSlug, cfg);
-      // The CRITICAL warning from findExamplesSource must be present.
-      expect(
-        a.warnings.some(
-          (w) => w.includes("unreadable-candidates") && w.includes(mappedSlug),
-        ),
-      ).toBe(true);
-      // No plain missing-examples anomaly.
-      expect(a.anomalies.some((x) => x.kind === "missing-examples")).toBe(
-        false,
-      );
-      // The new variant is present with the mapped candidates echoed.
-      const unreadable = a.anomalies.find(
-        (x) => x.kind === "unreadable-examples",
-      );
-      expect(unreadable).toBeDefined();
-      if (unreadable && unreadable.kind === "unreadable-examples") {
-        expect(unreadable.slug).toBe(mappedSlug);
-        expect([...unreadable.candidates]).toEqual([...candidates]);
-      }
-      // Rendering routes through anomalyMessage without throwing and
-      // carries the slug + candidates in the human-readable output.
-      const messages = anomalyStrings(a);
-      expect(
-        messages.some(
-          (m) =>
-            m.includes(mappedSlug) && candidates.every((c) => m.includes(c)),
-        ),
-      ).toBe(true);
-    } finally {
-      spy.mockRestore();
-    }
-  });
-
-  it("classifies unreadable-examples via structured metadata, not warning-string substring match (FX30-C Fix 1)", () => {
-    // Regression for FX30-C Fix 1 (R29-2 H1): auditPackage used to
-    // classify by scanning warnings[] for `unreadable-candidates` +
-    // `"<slug>"` substrings. Any reword of the human-readable warning
-    // text (for i18n, for a typo fix, for a docstring tweak) would
-    // silently reclassify unreadable-examples as missing-examples. The
-    // fix: route classification through a structured return value from
-    // resolveExamplesSource, not a string scan.
-    //
-    // Test strategy: confirm the contract via resolveExamplesSource's
-    // direct return shape (no warning string involved). The function
-    // must return a tagged object carrying `source` AND a boolean
-    // indicator that "all candidates existed but were unreadable" —
-    // tests below read that field directly.
-    const slug = "structural-unreadable";
-    const mapped = ["cand-a", "cand-b"] as const;
-    for (const c of mapped) makeExampleDir(root, c);
-    const fullPaths = mapped.map((c) =>
-      path.join(root, "examples", "integrations", c),
-    );
-    const orig = fs.statSync;
-    const spy = vi.spyOn(fs, "statSync").mockImplementation(((
-      p: fs.PathLike,
-      options?: unknown,
-    ) => {
-      if (typeof p === "string" && fullPaths.includes(p)) {
-        const e: NodeJS.ErrnoException = new Error("EACCES: injected");
-        e.code = "EACCES";
-        throw e;
-      }
-      return (orig as unknown as (p: fs.PathLike, o?: unknown) => unknown)(
-        p,
-        options,
-      );
-    }) as typeof fs.statSync);
-    try {
-      const cfg = makeConfig(root);
-      // resolveExamplesSource now returns a structured tuple, not a
-      // bare string. Assert both fields so the classification signal
-      // does NOT depend on warning-text substring matching.
-      const r = resolveExamplesSource(slug, mapped, cfg);
-      // The "all candidates unreadable" signal is a structured boolean.
-      expect(r).toMatchObject({
-        source: null,
-        unreadableForSlug: true,
-      });
-    } finally {
-      spy.mockRestore();
-    }
-  });
-
-  it("resolveExamplesSource returns unreadableForSlug=false for a stale mapping (no unreadable candidates)", () => {
-    // Negative control for the structured return: a stale mapping
-    // (candidates absent on disk) must NOT set unreadableForSlug=true.
-    // Otherwise audit would route plain stale-mapping cases as
-    // unreadable-examples (infrastructure failure) instead of the
-    // correct missing-examples (content failure).
-    const cfg = makeConfig(root);
-    const r = resolveExamplesSource(
-      "stale-mapping-slug",
-      ["absent-a", "absent-b"] as const,
-      cfg,
-    );
-    expect(r).toMatchObject({
-      source: null,
-      unreadableForSlug: false,
-    });
-  });
-
-  it("resolveExamplesSource returns the source string alongside unreadableForSlug=false on success", () => {
-    // On the happy path, `source` carries the relative path to the
-    // resolved candidate directory and `unreadableForSlug` is false.
-    makeExampleDir(root, "cand-hit");
-    const cfg = makeConfig(root);
-    const r = resolveExamplesSource(
-      "synthetic-hit",
-      ["cand-hit"] as const,
-      cfg,
-    );
-    expect(r).toMatchObject({
-      source: path.join("examples", "integrations", "cand-hit"),
-      unreadableForSlug: false,
-    });
-  });
-
-  it("auditPackage still routes unreadable-examples correctly even if the human-readable warning text is reworded", () => {
-    // Fix 1 robustness test: stub resolveExamplesSource's warning sink
-    // by monkey-patching the audit to emit a DIFFERENT warning wording,
-    // and confirm classification still lands on unreadable-examples. We
-    // achieve this by using statSync EACCES on the candidates (so the
-    // structured signal fires) and asserting classification does NOT
-    // require the human warning to contain `unreadable-candidates` or
-    // `"<slug>"` substrings verbatim.
-    const slug = "mastra";
-    const candidates = SLUG_TO_EXAMPLES[slug];
-    expect(candidates).toBeDefined();
-    writePackage(root, slug, {
-      manifest: `slug: ${slug}\ndeployed: true\ndemos:\n  - id: x\n`,
-      specs: ["x.spec.ts"],
-      qaFiles: ["x.md"],
-    });
-    const fullPaths = candidates.map((c) =>
-      path.join(root, "examples", "integrations", c),
-    );
-    for (const c of candidates) makeExampleDir(root, c);
-    const orig = fs.statSync;
-    const spy = vi.spyOn(fs, "statSync").mockImplementation(((
-      p: fs.PathLike,
-      options?: unknown,
-    ) => {
-      if (typeof p === "string" && fullPaths.includes(p)) {
-        const e: NodeJS.ErrnoException = new Error("EACCES: injected");
-        e.code = "EACCES";
-        throw e;
-      }
-      return (orig as unknown as (p: fs.PathLike, o?: unknown) => unknown)(
-        p,
-        options,
-      );
-    }) as typeof fs.statSync);
-    try {
-      const cfg = makeConfig(root);
-      const a = auditPackage(slug, cfg);
-      // Classification lands on unreadable-examples regardless of
-      // warning text. Fix 1 removed the string-substring gating.
-      expect(a.anomalies.some((x) => x.kind === "unreadable-examples")).toBe(
-        true,
-      );
-      expect(a.anomalies.some((x) => x.kind === "missing-examples")).toBe(
-        false,
-      );
-    } finally {
-      spy.mockRestore();
-    }
-  });
-
-  it("still pushes missing-examples (not unreadable-examples) when the mapping is simply stale", () => {
-    // Negative control for the branch above: the same mapped slug with
-    // NO candidate directories on disk must continue to produce
-    // missing-examples — the "unreadable-candidates" warning is
-    // specifically gated on candidates existing-but-unreadable.
-    const mappedSlug = "mastra";
-    writePackage(root, mappedSlug, {
-      manifest: `slug: ${mappedSlug}\ndeployed: true\ndemos:\n  - id: x\n`,
-      specs: ["x.spec.ts"],
-      qaFiles: ["x.md"],
-    });
-    // NOTE: no makeExampleDir — the mapping is stale.
-    const cfg = makeConfig(root);
-    const a = auditPackage(mappedSlug, cfg);
-    expect(a.anomalies.some((x) => x.kind === "missing-examples")).toBe(true);
-    expect(a.anomalies.some((x) => x.kind === "unreadable-examples")).toBe(
-      false,
-    );
-  });
-
-  it("ENOENT race between existsSync and statSync does NOT produce unreadable-examples (classified as missing-examples)", () => {
-    // TOCTOU race guard: existsSync can report true and a subsequent
-    // statSync can still throw ENOENT (the candidate directory was
-    // removed between the two calls, or we're on a network filesystem
-    // with weak coherence). ENOENT specifically indicates the candidate
-    // is simply NOT there by the time we stat it — that's a benign
-    // stale-mapping / disappearing-dir case, NOT an infrastructure
-    // failure. The resolver must NOT increment the unreadable tally
-    // for ENOENT, and auditPackage must NOT emit an `unreadable-examples`
-    // ERROR anomaly for this race. Fall through as if existsSync had
-    // returned false (→ `missing-examples` for mapped slugs).
-    const mappedSlug = "mastra";
-    const candidates = SLUG_TO_EXAMPLES[mappedSlug];
-    expect(candidates).toBeDefined();
-    writePackage(root, mappedSlug, {
-      manifest: `slug: ${mappedSlug}\ndeployed: true\ndemos:\n  - id: x\n`,
-      specs: ["x.spec.ts"],
-      qaFiles: ["x.md"],
-    });
-    for (const c of candidates) makeExampleDir(root, c);
-    const fullPaths = candidates.map((c) =>
-      path.join(root, "examples", "integrations", c),
-    );
-    const orig = fs.statSync;
-    const spy = vi.spyOn(fs, "statSync").mockImplementation(((
-      p: fs.PathLike,
-      options?: unknown,
-    ) => {
-      if (typeof p === "string" && fullPaths.includes(p)) {
-        const e: NodeJS.ErrnoException = new Error(
-          "ENOENT: no such file or directory",
-        );
-        e.code = "ENOENT";
-        throw e;
-      }
-      return (orig as unknown as (p: fs.PathLike, o?: unknown) => unknown)(
-        p,
-        options,
-      );
-    }) as typeof fs.statSync);
-    try {
-      const cfg = makeConfig(root);
-      const a = auditPackage(mappedSlug, cfg);
-      // Classification: ENOENT is a soft miss (absent candidate), not
-      // infrastructure. The resolver no longer differentiates "raced"
-      // from "never existed" because the existsSync pre-check is gone
-      // (existsSync conflates EACCES/ENOENT and was the whole source of
-      // the bugs this module fixes). ENOENT → continue, no warning.
-      expect(a.anomalies.some((x) => x.kind === "unreadable-examples")).toBe(
-        false,
-      );
-      expect(a.anomalies.some((x) => x.kind === "missing-examples")).toBe(true);
-      // No ERROR "unreadable-candidates" wording — that's reserved for
-      // true infrastructure failures (EACCES/EIO/etc.).
-      expect(a.warnings.some((w) => w.includes("unreadable-candidates"))).toBe(
-        false,
-      );
-      // The stale-mapping warning (no matching directory) is still the
-      // operator signal for this scenario.
-      expect(a.warnings.some((w) => w.includes("SLUG_TO_EXAMPLES entry"))).toBe(
-        true,
-      );
-    } finally {
-      spy.mockRestore();
-    }
-  });
-
-  it("resolveExamplesSource: ENOENT on statSync does NOT flip unreadableForSlug (soft miss)", () => {
-    // Unit-level guard for the same ENOENT branch. Purely structural:
-    // existedCount++ fires because existsSync returned true, but the
-    // ENOENT-specific catch path must NOT increment unreadableCount.
-    // Therefore unreadableForSlug stays false and the sink carries a
-    // diagnostic.
-    const slug = "synthetic-enoent";
-    const mapped = ["cand-raced"] as const;
-    makeExampleDir(root, "cand-raced");
-    const full = path.join(root, "examples", "integrations", "cand-raced");
-    const orig = fs.statSync;
-    const spy = vi.spyOn(fs, "statSync").mockImplementation(((
-      p: fs.PathLike,
-      options?: unknown,
-    ) => {
-      if (typeof p === "string" && p === full) {
-        const e: NodeJS.ErrnoException = new Error("ENOENT: raced");
-        e.code = "ENOENT";
-        throw e;
-      }
-      return (orig as unknown as (p: fs.PathLike, o?: unknown) => unknown)(
-        p,
-        options,
-      );
-    }) as typeof fs.statSync);
-    try {
-      const cfg = makeConfig(root);
-      const sink: string[] = [];
-      const r = resolveExamplesSource(slug, mapped, cfg, sink);
-      expect(r.source).toBeNull();
-      // Key assertion: ENOENT is a benign absence (soft miss), NOT
-      // infrastructure failure. With existsSync removed from the
-      // candidate gate, ENOENT from statSync is now indistinguishable
-      // from "never existed" — both route as missing-examples without
-      // a warning. The stale-mapping sink entry from the end of
-      // resolveExamplesSource (no matching directory) is the operator
-      // signal.
-      expect(r.unreadableForSlug).toBe(false);
-      // Stale-mapping diagnostic at the end of resolveExamplesSource
-      // fires because the mapped slug hit no successful dir.
-      expect(sink.some((w) => w.includes("SLUG_TO_EXAMPLES entry"))).toBe(true);
-    } finally {
-      spy.mockRestore();
-    }
-  });
-
-  it("mapped slug whose candidate exists but is NOT a directory surfaces a distinct anomaly (not silent missing-examples)", () => {
-    // For a MAPPED slug whose candidate path exists-but-is-not-a-
-    // directory (regular file / symlink-to-file / socket / FIFO), the
-    // old code existedCount++'d but unreadableCount stayed 0, so the
-    // slug silently fell through to `missing-examples` with NO
-    // diagnostic. Operators had no signal that the integrations dir had
-    // a stray file masquerading as the provenance target. The fix adds
-    // a distinct `mapped-candidate-not-directory` Anomaly kind driven
-    // by a structured `nonDirectoryForSlug` signal — never a warning-
-    // string substring match.
-    const slug = "mastra";
-    const candidates = SLUG_TO_EXAMPLES[slug];
-    expect(candidates).toBeDefined();
-    expect(candidates.length).toBeGreaterThan(0);
-    writePackage(root, slug, {
-      manifest: `slug: ${slug}\ndeployed: true\ndemos:\n  - id: x\n`,
-      specs: ["x.spec.ts"],
-      qaFiles: ["x.md"],
-    });
-    // Write a regular file (not a directory) at every candidate path.
-    for (const c of candidates) {
-      const p = path.join(root, "examples", "integrations", c);
-      fs.writeFileSync(p, "stray file, not a directory\n");
-    }
-    const cfg = makeConfig(root);
-    const a = auditPackage(slug, cfg);
-    // The new distinct anomaly kind must fire.
-    const notDir = a.anomalies.find(
-      (x) => x.kind === "mapped-candidate-not-directory",
-    );
-    expect(
-      notDir,
-      `expected mapped-candidate-not-directory anomaly, got: ${JSON.stringify(a.anomalies)}`,
-    ).toBeDefined();
-    if (notDir && notDir.kind === "mapped-candidate-not-directory") {
-      expect(notDir.slug).toBe(slug);
-      expect([...notDir.candidates]).toEqual([...candidates]);
-    }
-    // And MUST NOT silently degrade to missing-examples — the whole
-    // point of the fix is to make this misconfiguration visible.
-    expect(a.anomalies.some((x) => x.kind === "missing-examples")).toBe(false);
-    // Rendering routes through anomalyMessage without throwing and
-    // carries the slug + candidates in human-readable output.
-    const messages = anomalyStrings(a);
-    expect(
-      messages.some(
-        (m) => m.includes(slug) && candidates.every((c) => m.includes(c)),
-      ),
-    ).toBe(true);
-  });
-
-  it("mapped slug with mixed non-directory + missing candidates still emits mapped-candidate-not-directory", () => {
-    // Variant: one candidate is a stray file, another is simply absent.
-    // Since at least one mapped candidate existed-but-wasn't-a-dir, the
-    // distinct signal must still fire (not silent missing-examples).
-    const slug = "synthetic-mixed";
-    const mapped = ["stray-file", "totally-absent"] as const;
-    // Only create a regular file at the first candidate. The second
-    // candidate does not exist at all.
-    fs.writeFileSync(
-      path.join(root, "examples", "integrations", "stray-file"),
-      "not a dir\n",
-    );
-    const cfg = makeConfig(root);
-    const sink: string[] = [];
-    const r = resolveExamplesSource(slug, mapped, cfg, sink);
-    expect(r.source).toBeNull();
-    // Structural contract: resolver must communicate "candidate existed
-    // but was not a directory" separately from "all unreadable".
-    // unreadableForSlug stays false (no I/O failure). A new structured
-    // signal drives the mapped-candidate-not-directory anomaly.
-    expect(r.unreadableForSlug).toBe(false);
-    expect(r.nonDirectoryForSlug).toBe(true);
-    // A diagnostic warning must be present so operators see the
-    // misconfiguration.
-    expect(
-      sink.some(
-        (w) =>
-          w.includes("stray-file") && /not a directory|non-directory/i.test(w),
-      ),
-      `expected non-directory warning for stray-file in sink: ${JSON.stringify(sink)}`,
-    ).toBe(true);
-  });
-});
-
-describe("buildReport", () => {
-  let root: string;
-  beforeEach(() => {
-    root = makeTmpTree();
-  });
-  afterEach(() => {
-    fs.rmSync(root, { recursive: true, force: true });
-  });
-
-  it("counts clean vs anomaly packages correctly", () => {
-    writePackage(root, "crewai-crews", {
-      manifest: `slug: crewai-crews\ndeployed: true\ndemos:\n  - id: a\n`,
-      specs: ["a.spec.ts"],
-      qaFiles: ["a.md"],
-    });
-    makeExampleDir(root, "crewai-crews");
-    writePackage(root, "ag2", {
-      manifest: `slug: ag2\ndeployed: true\ndemos:\n  - id: a\n`,
-      specs: ["a.spec.ts"],
-      qaFiles: ["a.md"],
-    });
-    writePackage(root, "anomalous", {
-      manifest: `slug: anomalous\ndeployed: false\ndemos:\n  - id: a\n  - id: b\n`,
-      specs: ["a.spec.ts"],
-      qaFiles: ["a.md"],
-    });
-    const cfg = makeConfig(root);
-    const report = buildReport(["crewai-crews", "ag2", "anomalous"], cfg);
-    expect(report.totals.total).toBe(3);
-    expect(report.totals.clean).toBe(2);
-    expect(report.totals.withAnomalies).toBe(1);
-  });
-
-  it("classifies missing vs malformed manifests into distinct buckets", () => {
-    writePackage(root, "missing1", {}); // no manifest.yaml at all
-    writePackage(root, "missing2", {});
-    writePackage(root, "bad", { manifest: "demos: [[[\nunterminated\n" });
-    const cfg = makeConfig(root);
-    const report = buildReport(["bad", "missing1", "missing2"], cfg);
-    expect([...report.anomalies.missingManifest].sort()).toEqual(
-      ["missing1", "missing2"].sort(),
-    );
-    expect(report.anomalies.malformedManifest).toEqual(["bad"]);
-    // Cross-check: no overlap between the two buckets.
-    const overlap = report.anomalies.missingManifest.filter((s) =>
-      report.anomalies.malformedManifest.includes(s),
-    );
-    expect(overlap).toEqual([]);
-  });
-
-  it("per-dimension countMismatches filter: unreadable spec + real qa mismatch still appears", () => {
-    // Package 'mixed' has an unreadable spec dir AND a genuine qa
-    // mismatch. The old filter would hide the whole package from
-    // countMismatches because *any* 'could not read' was treated as a
-    // reason to suppress. The fix: per-dimension suppression — qa
-    // mismatch remains visible.
-    writePackage(root, "mixed", {
-      manifest: `slug: mixed\ndeployed: true\ndemos:\n  - id: a\n  - id: b\n`,
-      specs: ["a.spec.ts", "b.spec.ts"],
-      qaFiles: ["only-one.md"], // 1 qa vs 2 demos — real mismatch
-    });
-    makeExampleDir(root, "mixed");
-
-    const e2eDir = path.join(root, "packages", "mixed", "tests", "e2e");
-    const orig = fs.readdirSync;
-    const spy = vi.spyOn(fs, "readdirSync").mockImplementation(((
-      p: fs.PathLike,
-      options?: unknown,
-    ) => {
-      if (typeof p === "string" && p === e2eDir) {
-        const e: NodeJS.ErrnoException = new Error("EACCES");
-        e.code = "EACCES";
-        throw e;
-      }
-      return (orig as unknown as (p: fs.PathLike, o?: unknown) => unknown)(
-        p,
-        options,
-      );
-    }) as typeof fs.readdirSync);
-    try {
-      const cfg = makeConfig(root);
-      const report = buildReport(["mixed"], cfg);
-      // qa mismatch survives suppression (spec dir unreadable should
-      // only mask spec-dimension mismatches, not qa-dimension).
-      expect(report.anomalies.countMismatches).toContain("mixed");
-      // Also verify the underlying anomaly list — qa mismatch present,
-      // spec mismatch suppressed by "unreadable spec dir".
-      const audit = report.packages.find((p) => p.slug === "mixed")!;
-      const mismatches = audit.anomalies.filter(
-        (x): x is Extract<Anomaly, { kind: "count-mismatch" }> =>
-          x.kind === "count-mismatch",
-      );
-      expect(mismatches.some((m) => m.dimension === "qa")).toBe(true);
-      expect(mismatches.some((m) => m.dimension === "spec")).toBe(false);
-    } finally {
-      spy.mockRestore();
-    }
-  });
-
-  it("populates top-level hasAnomalies and exitCode", () => {
-    // JSON output needs the scalar summary so consumers don't have to
-    // re-derive it from nested arrays.
-    writePackage(root, "ok1", {
-      manifest: `slug: ok1\ndeployed: true\ndemos:\n  - id: a\n`,
-      specs: ["a.spec.ts"],
-      qaFiles: ["a.md"],
-    });
-    makeExampleDir(root, "ok1");
-    const cfg = makeConfig(root);
-    const clean = buildReport(["ok1"], cfg);
-    expect(clean.hasAnomalies).toBe(false);
-    expect(clean.exitCode).toBe(0);
-
-    writePackage(root, "bad", {
-      manifest: `slug: bad\ndeployed: false\ndemos:\n  - id: a\n`,
-      specs: [],
-      qaFiles: [],
-    });
-    const dirty = buildReport(["bad"], cfg);
-    expect(dirty.hasAnomalies).toBe(true);
-    expect(dirty.exitCode).toBe(1);
-  });
-
-  it("withAnomalies is a unique-package count even when buckets overlap", () => {
-    // Package "multi" has BOTH a count mismatch AND not-deployed. It
-    // appears in countMismatches AND notDeployed (buckets overlap), but
-    // totals.withAnomalies counts it once.
-    writePackage(root, "multi", {
-      manifest: `slug: multi\ndeployed: false\ndemos:\n  - id: a\n  - id: b\n`,
-      specs: ["a.spec.ts"], // 1 vs 2 — real spec mismatch
-      qaFiles: ["a.md", "b.md"],
-    });
-    makeExampleDir(root, "multi");
-    const cfg = makeConfig(root);
-    const report = buildReport(["multi"], cfg);
-    expect(report.anomalies.countMismatches).toContain("multi");
-    expect(report.anomalies.notDeployed).toContain("multi");
-    expect(report.totals.withAnomalies).toBe(1);
-  });
-
-  it("freezes PackageAudit entries to prevent downstream mutation", () => {
-    writePackage(root, "frozen", {
-      manifest: `slug: frozen\ndeployed: true\ndemos:\n  - id: a\n`,
-      specs: ["a.spec.ts"],
-      qaFiles: ["a.md"],
-    });
-    makeExampleDir(root, "frozen");
-    const cfg = makeConfig(root);
-    const report = buildReport(["frozen"], cfg);
-    const p = report.packages[0];
-    expect(Object.isFrozen(p)).toBe(true);
-  });
-
-  it("routes mapped-candidate-not-directory anomalies to a visible bucket (not invisible)", () => {
-    // Regression guard: the `mapped-candidate-not-directory` Anomaly
-    // variant is emitted by auditPackage for mapped slugs whose
-    // candidate path exists-but-isn't-a-directory. The bucket routing
-    // in buildReport previously had no case for this variant, so a
-    // package whose ONLY anomaly was `mapped-candidate-not-directory`
-    // appeared in `totals.withAnomalies` (exit 1) but was invisible in
-    // every bucket — renderAnomalySection had no section for it, so
-    // operators saw exit 1 with no explanation.
-    //
-    // Fix: route `mapped-candidate-not-directory` into the `unreadable`
-    // bucket (misconfiguration is closer to unreadable than to missing).
-    const slug = "mastra";
-    const candidates = SLUG_TO_EXAMPLES[slug];
-    expect(candidates).toBeDefined();
-    expect(candidates.length).toBeGreaterThan(0);
-    writePackage(root, slug, {
-      manifest: `slug: ${slug}\ndeployed: true\ndemos:\n  - id: x\n`,
-      specs: ["x.spec.ts"],
-      qaFiles: ["x.md"],
-    });
-    // Write a regular file (not a directory) at every candidate path so
-    // `mapped-candidate-not-directory` is the SOLE anomaly on this pkg.
-    for (const c of candidates) {
-      fs.writeFileSync(
-        path.join(root, "examples", "integrations", c),
-        "stray file, not a directory\n",
-      );
-    }
-    const cfg = makeConfig(root);
-    const report = buildReport([slug], cfg);
-    // The package surfaces ONE anomaly: mapped-candidate-not-directory.
-    const pkg = report.packages.find((p) => p.slug === slug)!;
-    expect(
-      pkg.anomalies.some((a) => a.kind === "mapped-candidate-not-directory"),
-    ).toBe(true);
-    // withAnomalies must count this package — which it already does.
-    expect(report.totals.withAnomalies).toBe(1);
-    // Critical: this package MUST show up in at least one indexed bucket
-    // so renderAnomalySection has something to render. We route it into
-    // `unreadable` (misconfiguration closer to unreadable than missing).
-    expect(report.anomalies.unreadable).toContain(slug);
-    // Negative: it must NOT be bucketed as missing-examples (that was
-    // the old silent-degradation behavior; the distinct variant is the
-    // whole point).
-    expect(report.anomalies.missingExamples).not.toContain(slug);
-  });
-
-  it("renderAnomalySection surfaces mapped-candidate-not-directory packages (not hidden)", async () => {
-    // Companion to the bucket-routing test: the text report must
-    // actually mention a package whose only anomaly is
-    // mapped-candidate-not-directory. Previously, such a package hit
-    // withAnomalies=1 but the anomaly section showed "(none)" because
-    // no bucket matched — an infuriating "CI failed but I don't know
-    // why" state.
-    const slug = "mastra";
-    const candidates = SLUG_TO_EXAMPLES[slug];
-    writePackage(root, slug, {
-      manifest: `slug: ${slug}\ndeployed: true\ndemos:\n  - id: x\n`,
-      specs: ["x.spec.ts"],
-      qaFiles: ["x.md"],
-    });
-    for (const c of candidates) {
-      fs.writeFileSync(
-        path.join(root, "examples", "integrations", c),
-        "stray file, not a directory\n",
-      );
-    }
-    const r = spawnSync("npx", ["tsx", AUDIT_SCRIPT], {
-      env: { ...process.env, SHOWCASE_AUDIT_ROOT: root },
-      encoding: "utf-8",
-      timeout: 30_000,
-    });
-    expect(r.status, r.stdout + r.stderr).toBe(1);
-    // The anomaly section must NOT say "(none)" — a package was flagged.
-    const anomalySection = r.stdout.split("Coverage anomalies")[1] ?? "";
-    const healthSection = anomalySection.split("Overall health")[0] ?? "";
-    expect(healthSection).not.toMatch(/\(none\)/);
-    // The slug MUST appear in the anomaly section.
-    expect(healthSection).toContain(slug);
-  });
-
-  it("deep-freezes PackageAudit inner containers (anomalies, warnings, spec, qa, manifest)", () => {
-    // Regression guard: a shallow freeze would leave
-    // `p.anomalies.push(...)` working, so downstream consumers could
-    // silently mutate the audit state through a nested reference. The
-    // contract is deep-frozen-enough: the record AND its mutable inner
-    // containers cannot be reassigned OR pushed to.
-    writePackage(root, "deepfrozen", {
-      manifest: `slug: deepfrozen\ndeployed: true\ndemos:\n  - id: a\n`,
-      specs: ["a.spec.ts"],
-      qaFiles: ["a.md"],
-    });
-    makeExampleDir(root, "deepfrozen");
-    const cfg = makeConfig(root);
-    const report = buildReport(["deepfrozen"], cfg);
-    const p = report.packages[0];
-    expect(Object.isFrozen(p)).toBe(true);
-    expect(Object.isFrozen(p.anomalies)).toBe(true);
-    expect(Object.isFrozen(p.warnings)).toBe(true);
-    expect(Object.isFrozen(p.spec)).toBe(true);
-    expect(Object.isFrozen(p.qa)).toBe(true);
-    expect(Object.isFrozen(p.manifest)).toBe(true);
-    // "ok" manifest variant should have its inner .manifest object
-    // frozen too.
-    if (p.manifest.kind === "ok") {
-      expect(Object.isFrozen(p.manifest.manifest)).toBe(true);
-    }
-  });
-});
-
-describe("parseArgs", () => {
-  it("treats --slug as requiring a non-flag argument (not --json)", () => {
-    const r = parseArgs(["--slug", "--json"]);
-    expect(r.slug).toBeNull();
-    expect(r.json).toBe(true);
-    expect(r.errors.some((e) => e.includes("--slug"))).toBe(true);
-  });
-
-  it("treats --slug followed by a flag as missing value", () => {
-    const r = parseArgs(["--slug"]);
-    expect(r.slug).toBeNull();
-    expect(r.errors.some((e) => e.includes("--slug"))).toBe(true);
-  });
-
-  it("parses --slug <value> correctly", () => {
-    const r = parseArgs(["--slug", "mastra"]);
-    expect(r.slug).toBe("mastra");
-    expect(r.errors).toEqual([]);
-  });
-
-  it("parses --json and --slug together", () => {
-    const r = parseArgs(["--json", "--slug", "mastra"]);
-    expect(r.slug).toBe("mastra");
-    expect(r.json).toBe(true);
-    expect(r.errors).toEqual([]);
-  });
-
-  it("rejects duplicate --json rather than silently accepting last-wins", () => {
-    const r = parseArgs(["--json", "--json"]);
-    expect(r.json).toBe(true);
-    expect(r.errors.some((e) => /--json/.test(e))).toBe(true);
-  });
-
-  it("rejects duplicate --slug rather than silently taking last-wins", () => {
-    const r = parseArgs(["--slug", "a", "--slug", "b"]);
-    expect(r.errors.some((e) => /--slug/.test(e))).toBe(true);
-    // Error message should mention both candidate values so the user can
-    // tell which occurrences collided.
-    expect(r.errors.some((e) => e.includes("a") && e.includes("b"))).toBe(true);
-  });
-});
-
-describe("BORN_IN_SHOWCASE set", () => {
-  it("contains the 5 known born-in-showcase slugs", () => {
-    expect(BORN_IN_SHOWCASE.has("ag2")).toBe(true);
-    expect(BORN_IN_SHOWCASE.has("claude-sdk-python")).toBe(true);
-    expect(BORN_IN_SHOWCASE.has("claude-sdk-typescript")).toBe(true);
-    expect(BORN_IN_SHOWCASE.has("langroid")).toBe(true);
-    expect(BORN_IN_SHOWCASE.has("spring-ai")).toBe(true);
-  });
-
-  it("every BORN_IN_SHOWCASE slug has no counterpart directory under examples/integrations/ (fixture-synthesized invariant)", () => {
-    // Fixture-based version of the real-repo invariant: synthesize a
-    // tmpdir that mimics examples/integrations/ containing only the
-    // slugs that SHOULD be there (i.e. nothing from BORN_IN_SHOWCASE).
-    // This always runs in CI — no `if (!fs.existsSync) return;` bail.
-    //
-    // The invariant asserted: given a clean examples/integrations/
-    // tree that contains no BORN_IN_SHOWCASE slugs, findExamplesSource
-    // must return null for every BORN_IN_SHOWCASE member. If someone
-    // later adds a BORN_IN_SHOWCASE slug to SLUG_TO_EXAMPLES by mistake
-    // this test will fail via the second assertion.
-    const tmp = makeTmpTree();
-    try {
-      // Seed a handful of non-born slugs so the examples dir isn't empty.
-      makeExampleDir(tmp, "mastra");
-      makeExampleDir(tmp, "agno");
-      const cfg = makeConfig(tmp);
-      for (const slug of BORN_IN_SHOWCASE) {
-        // No dir created for this slug — findExamplesSource must return
-        // a null source and SLUG_TO_EXAMPLES must not carry a mapping
-        // for it.
-        const r = findExamplesSource(slug, cfg);
-        expect(
-          r.source,
-          `BORN_IN_SHOWCASE slug "${slug}" resolved to a non-null examples source — the maps are inconsistent`,
-        ).toBeNull();
-        expect(
-          SLUG_TO_EXAMPLES[slug],
-          `BORN_IN_SHOWCASE slug "${slug}" appears in SLUG_TO_EXAMPLES — remove one or the other`,
-        ).toBeUndefined();
-      }
-    } finally {
-      fs.rmSync(tmp, { recursive: true, force: true });
-    }
-  });
-});
-
-describe("SLUG_TO_EXAMPLES — dead entries removed", () => {
-  // These showcase slugs do not exist under showcase/packages/. Keeping
-  // them in the map produced phantom "no examples source" anomalies for
-  // nothing (and made the table/doc noise confusing). Each is a
-  // regression guard: if you add any of these back, you must create the
-  // corresponding showcase/packages/<slug>/ dir too.
-  it.each(["crewai-flows", "agent-spec-langgraph", "mcp-apps"])(
-    "does not include dead entry %s",
-    (slug) => {
-      expect(SLUG_TO_EXAMPLES[slug]).toBeUndefined();
-    },
-  );
-
-  it("every mapped target resolves when its examples/integrations/ dir is present (fixture-synthesized)", () => {
-    // Fixture-based dead-entry guard: synthesize an examples/integrations/
-    // tmpdir containing every target named in SLUG_TO_EXAMPLES, then
-    // assert findExamplesSource returns a non-null path for each slug.
-    // This exercises the positive resolution path deterministically and
-    // always runs in CI — no `if (!fs.existsSync) return;` bail.
-    //
-    // Dead-entry protection: if anyone adds a SLUG_TO_EXAMPLES entry
-    // pointing at a non-existent target and the real repo lacks that
-    // dir, audit would emit a phantom "no examples source" anomaly.
-    // The slug-map.test.ts real-repo invariant (separate file) still
-    // enforces the production tree; this test locks in the runtime
-    // resolution contract.
-    const tmp = makeTmpTree();
-    try {
-      for (const targets of Object.values(SLUG_TO_EXAMPLES)) {
-        for (const target of targets) {
-          makeExampleDir(tmp, target);
-        }
-      }
-      const cfg = makeConfig(tmp);
-      for (const [slug, targets] of Object.entries(SLUG_TO_EXAMPLES)) {
-        const r = findExamplesSource(slug, cfg);
-        expect(
-          r.source,
-          `SLUG_TO_EXAMPLES[${slug}] → [${targets.join(
-            ", ",
-          )}] failed to resolve against a fixture containing every target`,
-        ).not.toBeNull();
-      }
-    } finally {
-      fs.rmSync(tmp, { recursive: true, force: true });
-    }
-  });
-});
-
-describe("findExamplesSource — sink-based warnings (no global stderr)", () => {
-  let root: string;
-  beforeEach(() => {
-    root = makeTmpTree();
-  });
-  afterEach(() => {
-    fs.rmSync(root, { recursive: true, force: true });
-  });
-
-  it("populates the caller-supplied sink when a mapped slug has no matching dir", () => {
-    // Use an explicit slug known to be in SLUG_TO_EXAMPLES rather than
-    // first-key indexing, which would silently keep passing if the map
-    // changed order.
-    const mappedSlug = "mastra";
-    expect(SLUG_TO_EXAMPLES[mappedSlug]).toBeDefined();
-    const cfg = makeConfig(root);
-    const sink: string[] = [];
-    const r = findExamplesSource(mappedSlug, cfg, sink);
-    expect(r.source).toBeNull();
-    const joined = sink.join("\n");
-    expect(joined).toMatch(/warn/i);
-    expect(joined).toContain(mappedSlug);
-  });
-
-  it("does NOT populate the sink for an unmapped slug missing a dir (falls back to slug==dirname)", () => {
-    // A slug not in SLUG_TO_EXAMPLES falls back to [slug]. That's the
-    // "no mapping" case — not a dead entry — so no warning.
-    const cfg = makeConfig(root);
-    const sink: string[] = [];
-    const r = findExamplesSource("totally-unmapped-slug", cfg, sink);
-    expect(r.source).toBeNull();
-    expect(sink).toEqual([]);
-  });
-
-  it("does NOT write to process.stderr at all (hard guarantee — no global state)", () => {
-    // Regression guard for the stderr-monkey-patch contract removal:
-    // findExamplesSource is a pure function with respect to
-    // stdout/stderr. Callers pass an explicit sink or accept that
-    // warnings are discarded.
-    const stderrWrites: string[] = [];
-    const spy = vi.spyOn(process.stderr, "write").mockImplementation(((
-      chunk: string | Uint8Array,
-    ) => {
-      stderrWrites.push(
-        typeof chunk === "string" ? chunk : Buffer.from(chunk).toString(),
-      );
-      return true;
-    }) as typeof process.stderr.write);
-    try {
-      const mappedSlug = "mastra";
-      const cfg = makeConfig(root);
-      findExamplesSource(mappedSlug, cfg, []);
-      expect(stderrWrites).toEqual([]);
-    } finally {
-      spy.mockRestore();
-    }
-  });
-});
-
-describe("findExamplesSource — unreadable-candidates ERROR branch", () => {
-  // Coverage for audit.ts:429-434: when a mapped slug has multiple
-  // candidates and ALL of them exist but ALL statSync calls fail, the
-  // resolver must emit an ERROR warning (category:
-  // unreadable-candidates) and return null. This is materially
-  // different from the benign "no matching dir" warning because we
-  // can't actually tell whether the provenance is satisfied — the
-  // downstream consumer needs the ERROR level to route differently.
-  let root: string;
-  beforeEach(() => {
-    root = makeTmpTree();
-  });
-  afterEach(() => {
-    fs.rmSync(root, { recursive: true, force: true });
-  });
-
-  it("emits an ERROR (category: unreadable-candidates) when every mapped candidate exists but statSync fails on all of them", () => {
-    const slug = "synthetic-unreadable";
-    const mapped = ["cand-a", "cand-b"] as const;
-    makeExampleDir(root, "cand-a");
-    makeExampleDir(root, "cand-b");
-    const dirA = path.join(root, "examples", "integrations", "cand-a");
-    const dirB = path.join(root, "examples", "integrations", "cand-b");
-    const orig = fs.statSync;
-    const spy = vi.spyOn(fs, "statSync").mockImplementation(((
-      p: fs.PathLike,
-      options?: unknown,
-    ) => {
-      if (typeof p === "string" && (p === dirA || p === dirB)) {
-        const e: NodeJS.ErrnoException = new Error("EACCES: injected");
-        e.code = "EACCES";
-        throw e;
-      }
-      return (orig as unknown as (p: fs.PathLike, o?: unknown) => unknown)(
-        p,
-        options,
-      );
-    }) as typeof fs.statSync);
-    try {
-      const cfg = makeConfig(root);
-      const sink: string[] = [];
-      const r = resolveExamplesSource(slug, mapped, cfg, sink);
-      expect(r.source).toBeNull();
-      // Structured classification signal: "all candidates unreadable"
-      // must surface as unreadableForSlug=true without relying on the
-      // human-readable warning wording.
-      expect(r.unreadableForSlug).toBe(true);
-      // The ERROR warning must be present, tagged with the category,
-      // and must name the slug so downstream consumers can route it.
-      expect(
-        sink.some((w) => /ERROR/.test(w) && w.includes(slug)),
-        `expected an ERROR warning for slug "${slug}" in sink: ${JSON.stringify(
-          sink,
-        )}`,
-      ).toBe(true);
-      expect(sink.some((w) => w.includes("unreadable-candidates"))).toBe(true);
-      // Every candidate should have shown up in a statSync warning too
-      // (companion diagnostic emitted from the loop body).
-      expect(
-        sink.some((w) => w.includes("statSync") && w.includes("cand-a")),
-      ).toBe(true);
-      expect(
-        sink.some((w) => w.includes("statSync") && w.includes("cand-b")),
-      ).toBe(true);
-    } finally {
-      spy.mockRestore();
-    }
-  });
-});
-
-describe("isProgrammerBug classification", () => {
-  // Coverage for audit.ts:1206-1222. The top-level catch in main()
-  // uses isProgrammerBug to decide between "broken invariant, worth a
-  // bug report" diagnostic wording and "infrastructure / I/O" wording.
-  // Both land on EXIT_INTERNAL but the operator-facing message diverges.
-  it("classifies a bare TypeError as a programmer bug", () => {
-    expect(isProgrammerBug(new TypeError("invariant broken"))).toBe(true);
-  });
-
-  it("classifies ReferenceError and RangeError as programmer bugs too", () => {
-    expect(isProgrammerBug(new ReferenceError("x is not defined"))).toBe(true);
-    expect(isProgrammerBug(new RangeError("Maximum call stack"))).toBe(true);
-  });
-
-  it("classifies a plain Error (no .code) as NOT a programmer bug", () => {
-    // A bare Error without `.code` is neither an errno nor one of the
-    // recognised programmer-bug subclasses — falls through the guard
-    // and returns false.
-    expect(isProgrammerBug(new Error("something else"))).toBe(false);
-  });
-
-  it("classifies an ErrnoException-shaped Error as NOT a programmer bug", () => {
-    // Any Error carrying a string `.code` is treated as runtime I/O —
-    // including the rare case where a TypeError also picks up a
-    // `.code` field, in which case the errno check wins.
-    const e: NodeJS.ErrnoException = new Error("EACCES: permission denied");
-    e.code = "EACCES";
-    expect(isProgrammerBug(e)).toBe(false);
-  });
-
-  it("classifies a TypeError carrying a string .code as NOT a programmer bug (errno wins)", () => {
-    // This is the "weird subclass drift" case called out in the
-    // comment at audit.ts:1197-1202 — an errno-bearing TypeError is
-    // biased toward runtime, not programmer.
-    const e = new TypeError("weird drift") as TypeError & { code?: string };
-    e.code = "EIO";
-    expect(isProgrammerBug(e)).toBe(false);
-  });
-
-  it("classifies non-Error values as NOT programmer bugs", () => {
-    // Non-Error throws (strings, numbers, plain objects) don't satisfy
-    // the `instanceof Error` guard and return false.
-    expect(isProgrammerBug("string thrown")).toBe(false);
-    expect(isProgrammerBug(42)).toBe(false);
-    expect(isProgrammerBug({ message: "plain" })).toBe(false);
-    expect(isProgrammerBug(undefined)).toBe(false);
-    expect(isProgrammerBug(null)).toBe(false);
-  });
-});
-
-describe("main() exit codes via CLI subprocess", () => {
-  let root: string;
-
-  beforeEach(() => {
-    root = makeTmpTree();
-  });
-
-  afterEach(() => {
-    fs.rmSync(root, { recursive: true, force: true });
-  });
-
-  function runCli(
-    args: string[],
-    opts: { cwd?: string; env?: NodeJS.ProcessEnv } = {},
-  ) {
-    return spawnSync("npx", ["tsx", AUDIT_SCRIPT, ...args], {
-      cwd: opts.cwd,
-      env: { ...process.env, ...opts.env },
-      encoding: "utf-8",
-      timeout: 30_000,
-    });
-  }
-
-  it("exits 0 when there are no anomalies", () => {
-    writePackage(root, "crewai-crews", {
-      manifest: `slug: crewai-crews\ndeployed: true\ndemos:\n  - id: a\n`,
-      specs: ["a.spec.ts"],
-      qaFiles: ["a.md"],
-    });
-    makeExampleDir(root, "crewai-crews");
-    const r = runCli([], {
-      env: { SHOWCASE_AUDIT_ROOT: root },
-    });
-    expect(r.status, r.stdout + r.stderr).toBe(0);
-    // Positive stdout assertions: a regression that truncates stdout or
-    // drops the summary/table would slip past a r.status-only check.
-    // Pin the column headers as whole tokens, the Overall health
-    // counters, and the clean-state affirmation line.
-    expect(r.stdout).toMatch(/\bslug\b/);
-    expect(r.stdout).toMatch(/\bdemos\b/);
-    expect(r.stdout).toMatch(/\bspecs\b/);
-    expect(r.stdout).toMatch(/\bdeployed\b/);
-    expect(r.stdout).toMatch(/\bexamples src\b/);
-    expect(r.stdout).toMatch(/Packages total:\s+1/);
-    expect(r.stdout).toMatch(/Clean:\s+1/);
-    expect(r.stdout).toMatch(/With anomalies:\s+0/);
-    expect(r.stdout).toMatch(/All packages pass coverage audit/);
-    // The fixture slug must appear as its own row in the table.
-    expect(r.stdout).toMatch(/\bcrewai-crews\b/);
-  });
-
-  it("exits 1 when anomalies are found", () => {
-    writePackage(root, "bad", {
-      manifest: `slug: bad\ndeployed: false\ndemos:\n  - id: a\n`,
-      specs: [],
-      qaFiles: [],
-    });
-    const r = runCli([], {
-      env: { SHOWCASE_AUDIT_ROOT: root },
-    });
-    expect(r.status, r.stdout + r.stderr).toBe(1);
-    // Positive stdout assertions: the anomaly report must include the
-    // slug and at least one of the expected anomaly categories for the
-    // fixture (count mismatch, not deployed, missing examples).
-    expect(r.stdout).toMatch(/\bbad\b/);
-    expect(r.stdout).toMatch(/Coverage anomalies/);
-    expect(r.stdout).toMatch(/Packages total:\s+1/);
-    expect(r.stdout).toMatch(/With anomalies:\s+1/);
-  });
-
-  it("exits 3 (unreadable) when SHOWCASE_AUDIT_ROOT points to missing packages dir", () => {
-    // Missing/unreadable packages dir is infrastructure failure, not
-    // user-input failure — distinct exit code from "invalid content" (2).
-    const empty = fs.mkdtempSync(path.join(os.tmpdir(), "audit-empty-"));
-    try {
-      const r = runCli([], {
-        env: { SHOWCASE_AUDIT_ROOT: empty },
-      });
-      expect(r.status, r.stdout + r.stderr).toBe(3);
-      // Tighten from /packages/i — that would also match log lines that
-      // merely name the string "packages". Pin the specific diagnostic
-      // phrase so a regression that swallows the reason (and exits 3
-      // for some other cause) can't slip through.
-      expect(r.stderr).toMatch(/packages dir does not exist/);
-    } finally {
-      fs.rmSync(empty, { recursive: true, force: true });
-    }
-  });
-
-  it("exits 3 (unreadable) when packages path exists but is a file, not a directory", () => {
-    // Regression guard: previously `readdirSync` on a file path threw
-    // ENOTDIR inside the try/catch in listShowcasePackageSlugs which
-    // returned [], so the CLI collapsed to "empty packages" (exit 1). We
-    // now distinguish this with a dedicated stat() check — exit 3.
-    const fixture = fs.mkdtempSync(path.join(os.tmpdir(), "audit-file-"));
-    try {
-      // Create <fixture>/packages as a FILE, not a directory.
-      fs.writeFileSync(path.join(fixture, "packages"), "not a dir\n");
-      const r = runCli([], {
-        env: { SHOWCASE_AUDIT_ROOT: fixture },
-      });
-      expect(r.status, r.stdout + r.stderr).toBe(3);
-      expect(r.stderr).toMatch(/not a directory/i);
-      // Tighten: the diagnostic MUST NOT claim the path "does not exist"
-      // — the file is present, just not a directory. The previous
-      // redundant `existsSync` pre-check produced the misleading "does
-      // not exist" wording for this case. After the fix, the statSync
-      // block produces a precise "is not a directory" message.
-      expect(r.stderr).not.toMatch(/packages dir does not exist/);
-    } finally {
-      fs.rmSync(fixture, { recursive: true, force: true });
-    }
-  });
-
-  it("exits 3 with a precise EACCES diagnostic when packages dir stat fails (not the misleading 'does not exist' message)", () => {
-    // Regression guard: main()'s redundant `fs.existsSync` pre-check
-    // emitted "packages dir does not exist" for EACCES/EPERM/EIO failures
-    // too — `existsSync` returns false for every statSync failure, not
-    // just ENOENT. The fix removes the redundant pre-check so the
-    // subsequent try/statSync block produces an accurate errno-specific
-    // message.
-    //
-    // Inject an EACCES via a preload script that overrides fs.statSync
-    // only for the target packages dir path (everything else passes
-    // through to the real implementation so tsx/vitest internals keep
-    // working).
-    const fixture = fs.mkdtempSync(path.join(os.tmpdir(), "audit-eacces-"));
-    const preload = fs.mkdtempSync(path.join(os.tmpdir(), "audit-pre-"));
-    const preloadScript = path.join(preload, "eacces.cjs");
-    const pkgDir = path.join(fixture, "packages");
-    // Create the dir so existsSync would return true — the bug is that
-    // statSync failing with EACCES should yield a distinct message. The
-    // old redundant existsSync check short-circuits ENOENT only; EACCES
-    // still falls through to statSync. But we also want to assert the
-    // message doesn't claim "does not exist" — exercise the code path
-    // by making statSync throw EACCES directly.
-    fs.mkdirSync(pkgDir, { recursive: true });
-    fs.writeFileSync(
-      preloadScript,
-      `const fs = require("fs");
-const target = ${JSON.stringify(pkgDir)};
-const origStat = fs.statSync;
-fs.statSync = function(...args) {
-  if (String(args[0]) === target) {
-    const e = new Error("EACCES: permission denied, stat '" + target + "'");
-    e.code = "EACCES";
-    throw e;
-  }
-  return origStat.apply(this, args);
-};
-const origExists = fs.existsSync;
-fs.existsSync = function(...args) {
-  if (String(args[0]) === target) {
-    // Simulate existsSync hiding EACCES as "false" — the bug.
-    return false;
-  }
-  return origExists.apply(this, args);
-};
-`,
-    );
-    try {
-      const r = spawnSync(
-        "npx",
-        ["tsx", "--require", preloadScript, AUDIT_SCRIPT],
-        {
-          env: { ...process.env, SHOWCASE_AUDIT_ROOT: fixture },
-          encoding: "utf-8",
-          timeout: 30_000,
-        },
-      );
-      expect(r.status, r.stdout + r.stderr).toBe(3);
-      // The diagnostic MUST carry the EACCES errno (precise,
-      // actionable) rather than the misleading "does not exist" wording
-      // the redundant existsSync branch used to emit.
-      expect(r.stderr).toMatch(/EACCES/);
-      expect(r.stderr).not.toMatch(/packages dir does not exist/);
-    } finally {
-      fs.rmSync(fixture, { recursive: true, force: true });
-      fs.rmSync(preload, { recursive: true, force: true });
-    }
-  });
-
-  it("unreadable (3) and invalid-content (2) exit codes differ", () => {
-    // Regression guard: these two failure modes used to share exit code
-    // 2, which made it impossible for CI callers to distinguish
-    // "nothing to audit" from "I don't know what you meant".
-    const empty = fs.mkdtempSync(path.join(os.tmpdir(), "audit-diff-"));
-    try {
-      const unreadable = runCli([], {
-        env: { SHOWCASE_AUDIT_ROOT: empty },
-      });
-      const invalidArgs = runCli(["--slug", "--json"], {
-        env: { SHOWCASE_AUDIT_ROOT: empty },
-      });
-      expect(unreadable.status).not.toBe(invalidArgs.status);
-      expect(unreadable.status).toBe(3);
-      expect(invalidArgs.status).toBe(2);
-    } finally {
-      fs.rmSync(empty, { recursive: true, force: true });
-    }
-  });
-
-  it("exits 1 (anomaly) when packages dir exists but is empty", () => {
-    // tree already has empty packages dir from makeTmpTree
-    const r = runCli([], {
-      env: { SHOWCASE_AUDIT_ROOT: root },
-    });
-    expect(r.status, r.stdout + r.stderr).toBe(1);
-  });
-
-  it("exits 2 on invalid arg combination (bad arg: --slug --json)", () => {
-    writePackage(root, "crewai-crews", {
-      manifest: `slug: crewai-crews\ndeployed: true\ndemos:\n  - id: a\n`,
-      specs: ["a.spec.ts"],
-      qaFiles: ["a.md"],
-    });
-    makeExampleDir(root, "crewai-crews");
-    const r = runCli(["--slug", "--json"], {
-      env: { SHOWCASE_AUDIT_ROOT: root },
-    });
-    // argparse failure is a user/internal error, not a package anomaly.
-    expect(r.status, r.stdout + r.stderr).toBe(2);
-  });
-
-  it("--json --slug <slug> combination emits JSON for a single package", () => {
-    writePackage(root, "crewai-crews", {
-      manifest: `slug: crewai-crews\ndeployed: true\ndemos:\n  - id: a\n`,
-      specs: ["a.spec.ts"],
-      qaFiles: ["a.md"],
-    });
-    makeExampleDir(root, "crewai-crews");
-    writePackage(root, "other", {
-      manifest: `slug: other\ndeployed: true\ndemos:\n  - id: a\n`,
-      specs: ["a.spec.ts"],
-      qaFiles: ["a.md"],
-    });
-    makeExampleDir(root, "other");
-    const r = runCli(["--json", "--slug", "crewai-crews"], {
-      env: { SHOWCASE_AUDIT_ROOT: root },
-    });
-    expect(r.status, r.stdout + r.stderr).toBe(0);
-    const parsed = JSON.parse(r.stdout);
-    expect(parsed.packages.length).toBe(1);
-    expect(parsed.packages[0].slug).toBe("crewai-crews");
-    // Scalar summary exposed alongside the nested report.
-    expect(parsed.hasAnomalies).toBe(false);
-    expect(parsed.exitCode).toBe(0);
-  });
-
-  it("JSON mode does not duplicate per-package warnings to stderr", () => {
-    // In JSON mode, warnings are already carried on
-    // `packages[i].warnings` — echoing them to stderr would
-    // double-emit the same information. A consumer redirecting
-    // `2>/dev/null` should still get a complete machine-readable
-    // report via stdout.
-    const mappedSlug = "mastra";
-    expect(SLUG_TO_EXAMPLES[mappedSlug]).toBeDefined();
-    writePackage(root, mappedSlug, {
-      manifest: `slug: ${mappedSlug}\ndeployed: true\ndemos:\n  - id: a\n`,
-      specs: ["a.spec.ts"],
-      qaFiles: ["a.md"],
-    });
-    // Intentionally DO NOT create examples/integrations/<mapped> dir so
-    // findExamplesSource emits a stale-mapping warning.
-    const r = runCli(["--json"], {
-      env: { SHOWCASE_AUDIT_ROOT: root },
-    });
-    // exit is 0/1 depending on anomalies — the focus here is stderr
-    // contents, not exit code.
-    expect(r.stderr || "").not.toMatch(/audit: warning:/);
-    // The JSON stdout should still carry the warning on the package
-    // record so JSON consumers aren't blind.
-    const parsed = JSON.parse(r.stdout);
-    const p = parsed.packages.find(
-      (x: { slug: string }) => x.slug === mappedSlug,
-    );
-    expect(p).toBeDefined();
-    expect(p.warnings.length).toBeGreaterThan(0);
-  });
-
-  it("text mode forwards per-package warnings to stderr for human readers", () => {
-    // Counterpart: in text mode a terminal user watching stderr should
-    // still see the stale-mapping diagnostic — the sink-based warnings
-    // must be forwarded, not silently dropped.
-    const mappedSlug = "mastra";
-    expect(SLUG_TO_EXAMPLES[mappedSlug]).toBeDefined();
-    writePackage(root, mappedSlug, {
-      manifest: `slug: ${mappedSlug}\ndeployed: true\ndemos:\n  - id: a\n`,
-      specs: ["a.spec.ts"],
-      qaFiles: ["a.md"],
-    });
-    const r = runCli([], {
-      env: { SHOWCASE_AUDIT_ROOT: root },
-    });
-    // Tighten from bare /audit: warning:/ — that would also accept any
-    // unrelated warning. Pin:
-    //   - the "audit: warning:" prefix (routing)
-    //   - the specific slug ("mastra") that triggered it
-    //   - the SLUG_TO_EXAMPLES phrase that identifies this as the
-    //     stale-mapping diagnostic, not some other warning
-    expect(r.stderr).toMatch(/audit: warning:/);
-    expect(r.stderr).toMatch(
-      new RegExp(`audit: warning: SLUG_TO_EXAMPLES entry "${mappedSlug}"`),
-    );
-    expect(r.stderr).toMatch(/has no matching directory/);
-  });
-
-  it("exits 4 (internal error) on unexpected exceptions", () => {
-    // Inject a TypeError into the YAML parser — auditPackage delegates
-    // to parseManifest which calls yaml.parse inside a try/catch that
-    // maps to a `malformed` result. HOWEVER, if we monkey-patch
-    // `yaml.parse` to throw a TypeError AFTER parseManifest has already
-    // read the file, the catch block inside parseManifest catches that
-    // and returns `malformed`. The cleanest way to trigger EXIT_INTERNAL
-    // is to throw a non-Error value from a point that is NOT wrapped
-    // upstream.
-    //
-    // Concretely: override `fs.readFileSync` to throw a TypeError ONLY
-    // for the target manifest.yaml path. readFileSync in parseManifest
-    // IS wrapped — but the catch returns `{kind: "unreadable", error}`
-    // only if `e instanceof Error`. A TypeError IS an Error, so that
-    // catch activates. That means we can't trigger EXIT_INTERNAL via
-    // readFileSync either.
-    //
-    // The actual unwrapped paths are: buildReport's Object.freeze calls,
-    // anomaly bucket operations, renderTable, renderAnomalySection,
-    // renderHealthSection, JSON.stringify. The simplest wedge is to
-    // monkey-patch `JSON.stringify` to throw TypeError when called with
-    // the audit report, which happens only in --json mode. For the
-    // default (text) mode, inject a TypeError through renderTable by
-    // monkey-patching `String.prototype.padEnd` — but that's too
-    // invasive.
-    //
-    // Simpler: force the failure in `Object.freeze` via a Proxy wrapper
-    // around the anomaly array. Actually, the cleanest wedge: override
-    // `Array.prototype.filter` in the preload so the FIRST call after
-    // slug listing throws. But that breaks tsx too.
-    //
-    // Practical solution: use `--json` mode and intercept
-    // `JSON.stringify` to throw once, after the report is built.
-    const preload = fs.mkdtempSync(path.join(os.tmpdir(), "audit-preload-"));
-    const preloadScript = path.join(preload, "boom.cjs");
-    fs.writeFileSync(
-      preloadScript,
-      `const origStringify = JSON.stringify;
-JSON.stringify = function(value, ...rest) {
-  // Only fire when serializing the audit report (has the 'packages'
-  // array + 'anomalies' object shape). Other JSON.stringify callers
-  // (test runner, tsx internals, error formatting) still work.
-  if (
-    value &&
-    typeof value === "object" &&
-    Array.isArray(value.packages) &&
-    value.anomalies &&
-    typeof value.anomalies === "object" &&
-    "countMismatches" in value.anomalies
-  ) {
-    throw new TypeError("simulated bug: should never happen");
-  }
-  return origStringify.call(this, value, ...rest);
-};
-`,
-    );
-    writePackage(root, "foo", {
-      manifest: `slug: foo\ndeployed: true\ndemos:\n  - id: a\n`,
-      specs: ["a.spec.ts"],
-      qaFiles: ["a.md"],
-    });
-    try {
-      const r = spawnSync(
-        "npx",
-        ["tsx", "--require", preloadScript, AUDIT_SCRIPT, "--json"],
-        {
-          env: { ...process.env, SHOWCASE_AUDIT_ROOT: root },
-          encoding: "utf-8",
-          timeout: 30_000,
-        },
-      );
-      // The injected failure fires from JSON.stringify in main()'s
-      // --json branch, which is NOT wrapped in a local try/catch — it
-      // bubbles to the top-level catch, which must route programmer
-      // bugs (TypeError) to EXIT_INTERNAL (4).
-      expect(r.status, r.stdout + r.stderr).toBe(4);
-      // stderr should use the programmer-bug wording, not the generic
-      // "internal error" one.
-      expect(r.stderr).toMatch(/bug \(programmer error\)/);
-    } finally {
-      fs.rmSync(preload, { recursive: true, force: true });
-    }
-  });
-
-  it("exits 3 with a clear error when SHOWCASE_AUDIT_ROOT points to a nonexistent path", () => {
-    // Regression guard: previously an invalid SHOWCASE_AUDIT_ROOT was
-    // accepted silently and users got a confusing downstream error about
-    // the derived `<root>/packages` path. We now validate the env-var
-    // path itself and emit a clear message naming SHOWCASE_AUDIT_ROOT.
-    const missing = path.join(
-      os.tmpdir(),
-      `audit-nonexistent-${Date.now()}-${Math.random().toString(36).slice(2)}`,
-    );
-    // Paranoia: ensure it really doesn't exist (collision with a prior
-    // run would mask the fix).
-    expect(fs.existsSync(missing)).toBe(false);
-    const r = runCli([], {
-      env: { SHOWCASE_AUDIT_ROOT: missing },
-    });
-    expect(r.status, r.stdout + r.stderr).toBe(3);
-    expect(r.stderr).toMatch(/SHOWCASE_AUDIT_ROOT/);
-    expect(r.stderr).toMatch(/does not exist/i);
-    expect(r.stderr).toContain(missing);
-  });
-
-  it("exits 3 with a clear error when SHOWCASE_AUDIT_ROOT points to a file, not a directory", () => {
-    // Counterpart regression guard: a file-typed SHOWCASE_AUDIT_ROOT used
-    // to flow through to the packages-dir check and emit an unhelpful
-    // "packages dir does not exist: <file>/packages" message. Now the
-    // env-var validation layer catches it first with a precise diagnostic.
-    const filePath = path.join(
-      os.tmpdir(),
-      `audit-root-as-file-${Date.now()}-${Math.random().toString(36).slice(2)}`,
-    );
-    fs.writeFileSync(filePath, "not a dir\n");
-    try {
-      const r = runCli([], {
-        env: { SHOWCASE_AUDIT_ROOT: filePath },
-      });
-      expect(r.status, r.stdout + r.stderr).toBe(3);
-      expect(r.stderr).toMatch(/SHOWCASE_AUDIT_ROOT/);
-      expect(r.stderr).toMatch(/not a directory/i);
-      expect(r.stderr).toContain(filePath);
-    } finally {
-      fs.rmSync(filePath, { force: true });
-    }
-  });
-});
-
-describe("UnreadableDirError", () => {
-  it("carries the dir and a human-readable message", () => {
-    const cause = new Error("EACCES: permission denied");
-    const e = new UnreadableDirError("/tmp/packages", cause);
-    expect(e).toBeInstanceOf(Error);
-    expect(e.name).toBe("UnreadableDirError");
-    expect(e.dir).toBe("/tmp/packages");
-    expect(e.message).toMatch(/^could not read \/tmp\/packages: EACCES/);
-    // ES2022 cause chain preserves the original ErrnoException so
-    // callers can still reach `.code` / `.errno` / `.syscall`.
-    expect((e as Error & { cause?: unknown }).cause).toBe(cause);
-  });
-
-  it("is thrown by listShowcasePackageSlugs when readdirSync fails", () => {
-    // Verify the error is raised at the right control-flow point:
-    // listShowcasePackageSlugs catches fs failures and rethrows as
-    // UnreadableDirError carrying the exact packages dir it tried.
-    const root = fs.mkdtempSync(path.join(os.tmpdir(), "audit-unreadable-"));
-    try {
-      const cfg = {
-        packagesDir: path.join(root, "does-not-exist"),
-        examplesIntegrationsDir: path.join(root, "examples", "integrations"),
-        repoRoot: root,
-      };
-      let thrown: unknown = null;
-      try {
-        listShowcasePackageSlugs(cfg);
-      } catch (e) {
-        thrown = e;
-      }
-      expect(thrown).toBeInstanceOf(UnreadableDirError);
-      const e = thrown as UnreadableDirError;
-      expect(e.dir).toBe(cfg.packagesDir);
-      expect(e.message).toContain(cfg.packagesDir);
-    } finally {
-      fs.rmSync(root, { recursive: true, force: true });
-    }
-  });
-
-  it("preserves the non-Error cause via String() coercion in the message", () => {
-    // Defensive: an fs layer could, in principle, reject with a
-    // non-Error value. UnreadableDirError still produces a readable
-    // message via String(cause).
-    const e = new UnreadableDirError("/tmp/x", "bespoke-failure-token");
-    expect(e.message).toBe("could not read /tmp/x: bespoke-failure-token");
-  });
-
-  it("includes the errno .code in the rendered message when missing from cause.message", () => {
-    // Some callers/tests construct Errors that attach .code but do not
-    // embed the code in .message. UnreadableDirError surfaces the code
-    // in its rendered message so operators see e.g. EACCES immediately.
-    const cause = Object.assign(new Error("permission denied"), {
-      code: "EACCES",
-    });
-    const e = new UnreadableDirError("/x", cause);
-    expect(e.message).toBe("could not read /x: EACCES: permission denied");
-  });
-
-  it("does not double-prepend errno code when already in cause.message", () => {
-    // Node's fs errors embed the code in .message (e.g.
-    // "EACCES: permission denied, scandir ..."). Avoid "EACCES: EACCES:".
-    const cause = Object.assign(
-      new Error("EACCES: permission denied, scandir '/x'"),
-      { code: "EACCES" },
-    );
-    const e = new UnreadableDirError("/x", cause);
-    expect(e.message).toBe(
-      "could not read /x: EACCES: permission denied, scandir '/x'",
-    );
-    expect(e.message).not.toMatch(/EACCES: EACCES/);
-  });
-});
-
-describe("canonicalizeForIsMain", () => {
-  it("returns a canonical (realpath) path when the file exists", () => {
-    // Two distinct input strings that refer to the same underlying file
-    // (one absolute, one with a redundant `.` segment) must canonicalize
-    // to the same result. This proves the helper normalizes inputs
-    // rather than merely passing them through — a stronger invariant
-    // than `canonicalize(f) === realpathSync(f)`, which would be
-    // trivially true for any pass-through implementation.
-    const root = fs.mkdtempSync(path.join(os.tmpdir(), "audit-canon-"));
-    try {
-      const f = path.join(root, "a.txt");
-      fs.writeFileSync(f, "x");
-      // Same file, two different string forms: plain absolute vs.
-      // absolute with a redundant `./` segment injected in the middle.
-      // Concatenate directly rather than via path.join (which would
-      // collapse `.` eagerly) so the inputs are genuinely distinct at
-      // the string level before the helper sees them.
-      const plain = f;
-      const dotted = `${root}${path.sep}.${path.sep}a.txt`;
-      expect(plain).not.toBe(dotted); // sanity: inputs are distinct strings
-      const canonPlain = canonicalizeForIsMain(plain);
-      const canonDotted = canonicalizeForIsMain(dotted);
-      expect(canonPlain).toBe(canonDotted);
-      expect(canonPlain).toBe(fs.realpathSync(f));
-    } finally {
-      fs.rmSync(root, { recursive: true, force: true });
-    }
-  });
-
-  it("silently falls back to resolved path on ENOENT (no stderr noise)", () => {
-    // Synthetic argv[0] paths produced by some test harnesses don't
-    // exist on disk; ENOENT must stay quiet so CI logs aren't polluted.
-    const spy = vi
-      .spyOn(process.stderr, "write")
-      .mockImplementation(() => true);
-    try {
-      const missing = path.join(
-        os.tmpdir(),
-        "does-not-exist-xyz-" + Date.now(),
-      );
-      const out = canonicalizeForIsMain(missing);
-      expect(out).toBe(path.resolve(missing));
-      expect(spy).not.toHaveBeenCalled();
-    } finally {
-      spy.mockRestore();
-    }
-  });
-
-  it("terminates the process with EXIT_UNREADABLE on non-ENOENT realpath failure", () => {
-    // Non-ENOENT realpath failures (EACCES/ELOOP/EIO) indicate genuine
-    // filesystem-access problems that CI must surface distinctly from a
-    // clean run (exit 0) or drift (exit 1). The helper runs pre-main
-    // during the isMain guard, so `process.exitCode` would be clobbered
-    // by main()'s later `process.exitCode = report.exitCode` write —
-    // `process.exit(3)` is the only way to guarantee the signal
-    // survives. We spy on process.exit to observe the code without
-    // actually terminating the test runner.
-    // Filter by path so tsx / vitest internals that call realpathSync
-    // during module resolution or stack-trace symbolication fall
-    // through to the real implementation — only the synthetic test
-    // path triggers the synthesized failure.
-    const origRealpath = fs.realpathSync;
-    const realpathSpy = vi.spyOn(fs, "realpathSync").mockImplementation(((
-      p: fs.PathLike,
-      ...rest: unknown[]
-    ) => {
-      if (String(p) === "/some/path") {
-        throw Object.assign(new Error("I/O error"), { code: "EIO" });
-      }
-      return (origRealpath as (...a: unknown[]) => string)(p, ...rest);
-    }) as unknown as typeof fs.realpathSync);
-    const stderrSpy = vi
-      .spyOn(process.stderr, "write")
-      .mockImplementation(() => true);
-    const exitSpy = vi.spyOn(process, "exit").mockImplementation(((
-      _code?: number,
-    ) => {
-      // Throw a sentinel so the caller's control flow stops — real
-      // process.exit never returns, so code after it expects to be
-      // unreachable.
-      throw new Error(`__EXIT_CALLED__:${_code}`);
-    }) as unknown as typeof process.exit);
-    try {
-      expect(() => canonicalizeForIsMain("/some/path")).toThrow(
-        /__EXIT_CALLED__:3/,
-      );
-      expect(exitSpy).toHaveBeenCalledWith(3);
-    } finally {
-      realpathSpy.mockRestore();
-      stderrSpy.mockRestore();
-      exitSpy.mockRestore();
-    }
-  });
-
-  it("does NOT set process.exitCode on ENOENT fallback", () => {
-    // ENOENT is the benign synthetic-argv[0] case — must stay quiet
-    // and must NOT leak an exit-code elevation to the surrounding run.
-    // Filter by path so only the synthetic test path triggers ENOENT —
-    // tsx / vitest internals continue to use the real implementation.
-    const origRealpath = fs.realpathSync;
-    const realpathSpy = vi.spyOn(fs, "realpathSync").mockImplementation(((
-      p: fs.PathLike,
-      ...rest: unknown[]
-    ) => {
-      if (String(p) === "/some/missing/path") {
-        throw Object.assign(new Error("no such file"), { code: "ENOENT" });
-      }
-      return (origRealpath as (...a: unknown[]) => string)(p, ...rest);
-    }) as unknown as typeof fs.realpathSync);
-    const priorExitCode = process.exitCode;
-    process.exitCode = 0;
-    try {
-      canonicalizeForIsMain("/some/missing/path");
-      expect(process.exitCode).toBe(0);
-    } finally {
-      process.exitCode = priorExitCode;
-      realpathSpy.mockRestore();
-    }
-  });
-
-  it("warns to stderr for non-ENOENT failures (e.g. EACCES) before terminating the process", () => {
-    // Non-ENOENT realpath failures indicate real problems (permission,
-    // loop, I/O). The helper must emit an operator-facing diagnostic
-    // to stderr BEFORE calling process.exit so the cause is visible in
-    // the terminated run's logs — a silent process.exit would produce
-    // a bare exit code with no explanation.
-    // Filter by path so only the synthetic test path raises EACCES —
-    // tsx / vitest internals continue to use the real implementation.
-    const origRealpath = fs.realpathSync;
-    const realpathSpy = vi.spyOn(fs, "realpathSync").mockImplementation(((
-      p: fs.PathLike,
-      ...rest: unknown[]
-    ) => {
-      if (String(p) === "/some/path") {
-        throw Object.assign(new Error("permission denied"), {
-          code: "EACCES",
-        });
-      }
-      return (origRealpath as (...a: unknown[]) => string)(p, ...rest);
-    }) as unknown as typeof fs.realpathSync);
-    const stderrSpy = vi
-      .spyOn(process.stderr, "write")
-      .mockImplementation(() => true);
-    const exitSpy = vi.spyOn(process, "exit").mockImplementation(((
-      _code?: number,
-    ) => {
-      throw new Error(`__EXIT_CALLED__:${_code}`);
-    }) as unknown as typeof process.exit);
-    try {
-      expect(() => canonicalizeForIsMain("/some/path")).toThrow(
-        /__EXIT_CALLED__:3/,
-      );
-      expect(stderrSpy).toHaveBeenCalledTimes(1);
-      const msg = String(stderrSpy.mock.calls[0][0]);
-      expect(msg).toContain("[canonicalizeForIsMain] realpath failed for");
-      expect(msg).toContain("/some/path");
-      expect(msg).toContain("permission denied");
-    } finally {
-      realpathSpy.mockRestore();
-      stderrSpy.mockRestore();
-      exitSpy.mockRestore();
-    }
-  });
-
-  it("does not crash when a non-Error primitive is thrown by realpathSync", () => {
-    // Hardening: e instanceof Error guard prevents
-    // `(e as NodeJS.ErrnoException).code` from crashing on primitive
-    // throws ("string", number, plain object). The helper should treat
-    // these as non-ENOENT and terminate with EXIT_UNREADABLE.
-    // Filter by path so only the synthetic test path throws the raw
-    // primitive — tsx / vitest internals continue to use the real impl.
-    const origRealpath = fs.realpathSync;
-    const realpathSpy = vi.spyOn(fs, "realpathSync").mockImplementation(((
-      p: fs.PathLike,
-      ...rest: unknown[]
-    ) => {
-      if (String(p) === "/some/path") {
-        // Intentionally throw a non-Error primitive.
-        // eslint-disable-next-line @typescript-eslint/no-throw-literal
-        throw "raw string failure";
-      }
-      return (origRealpath as (...a: unknown[]) => string)(p, ...rest);
-    }) as unknown as typeof fs.realpathSync);
-    const stderrSpy = vi
-      .spyOn(process.stderr, "write")
-      .mockImplementation(() => true);
-    const exitSpy = vi.spyOn(process, "exit").mockImplementation(((
-      _code?: number,
-    ) => {
-      throw new Error(`__EXIT_CALLED__:${_code}`);
-    }) as unknown as typeof process.exit);
-    try {
-      expect(() => canonicalizeForIsMain("/some/path")).toThrow(
-        /__EXIT_CALLED__:3/,
-      );
-      expect(stderrSpy).toHaveBeenCalledTimes(1);
-      const msg = String(stderrSpy.mock.calls[0][0]);
-      expect(msg).toContain("raw string failure");
-    } finally {
-      realpathSpy.mockRestore();
-      stderrSpy.mockRestore();
-      exitSpy.mockRestore();
-    }
-  });
-});
-
-describe("listShowcasePackageSlugs", () => {
-  let root: string;
-  beforeEach(() => {
-    root = makeTmpTree();
-  });
-  afterEach(() => {
-    fs.rmSync(root, { recursive: true, force: true });
-  });
-
-  it("returns sorted slug list", () => {
-    writePackage(root, "beta", {});
-    writePackage(root, "alpha", {});
-    writePackage(root, "gamma", {});
-    const cfg = makeConfig(root);
-    expect(listShowcasePackageSlugs(cfg)).toEqual(["alpha", "beta", "gamma"]);
-  });
-
-  it("throws UnreadableDirError when packages dir cannot be read", () => {
-    const cfg = {
-      ...makeConfig(root),
-      packagesDir: path.join(root, "nope"),
-    };
-    // Missing dir → readdirSync ENOENT → UnreadableDirError. Previously
-    // the code returned [] and main() collapsed this to "exit 1 (empty
-    // packages)", masking a real I/O failure.
-    expect(() => listShowcasePackageSlugs(cfg)).toThrow(UnreadableDirError);
-  });
-});
-
-describe("auditPackage — direct-caller invariants", () => {
-  let root: string;
-  beforeEach(() => {
-    root = makeTmpTree();
-  });
-  afterEach(() => {
-    fs.rmSync(root, { recursive: true, force: true });
-  });
-
-  it("anomalies and warnings arrays are frozen on the audit returned by auditPackage (before buildReport)", () => {
-    // Regression guard: previously the arrays were only frozen inside
-    // buildReport, so direct callers of auditPackage saw unfrozen arrays
-    // and could mutate them.
-    writePackage(root, "direct", {
-      manifest: `slug: direct\ndeployed: true\ndemos:\n  - id: a\n`,
-      specs: ["a.spec.ts"],
-      qaFiles: ["a.md"],
-    });
-    makeExampleDir(root, "direct");
-    const cfg = makeConfig(root);
-    const a = auditPackage("direct", cfg);
-    expect(Object.isFrozen(a.anomalies)).toBe(true);
-    expect(Object.isFrozen(a.warnings)).toBe(true);
-  });
-
-  it("Anomaly.not-deployed.state distinguishes 'unset' from 'explicit-false'", () => {
-    writePackage(root, "unsetreal", {
-      manifest: `slug: unsetreal\ndemos:\n  - id: x\n`,
-      specs: ["x.spec.ts"],
-      qaFiles: ["x.md"],
-    });
-    makeExampleDir(root, "unsetreal");
-    writePackage(root, "falsereal", {
-      manifest: `slug: falsereal\ndeployed: false\ndemos:\n  - id: x\n`,
-      specs: ["x.spec.ts"],
-      qaFiles: ["x.md"],
-    });
-    makeExampleDir(root, "falsereal");
-    const cfg = makeConfig(root);
-    const a = auditPackage("unsetreal", cfg);
-    const b = auditPackage("falsereal", cfg);
-    const aNot = a.anomalies.find(
-      (x): x is Extract<Anomaly, { kind: "not-deployed" }> =>
-        x.kind === "not-deployed",
-    );
-    const bNot = b.anomalies.find(
-      (x): x is Extract<Anomaly, { kind: "not-deployed" }> =>
-        x.kind === "not-deployed",
-    );
-    expect(aNot).toBeDefined();
-    expect(bNot).toBeDefined();
-    // Self-documenting string union — callers read the raw boolean off
-    // the manifest variant when they need the original value.
-    expect(aNot!.state).toBe("unset");
-    expect(bNot!.state).toBe("explicit-false");
-  });
-
-  it("Object.freeze on manifest.manifest also freezes the demos array and each demo", () => {
-    writePackage(root, "frozendemos", {
-      manifest: `slug: frozendemos\ndeployed: true\ndemos:\n  - id: a\n  - id: b\n`,
-      specs: ["a.spec.ts", "b.spec.ts"],
-      qaFiles: ["a.md", "b.md"],
-    });
-    makeExampleDir(root, "frozendemos");
-    const cfg = makeConfig(root);
-    const report = buildReport(["frozendemos"], cfg);
-    const p = report.packages[0];
-    expect(p.manifest.kind).toBe("ok");
-    if (p.manifest.kind === "ok") {
-      const demos = p.manifest.manifest.demos!;
-      expect(Object.isFrozen(demos)).toBe(true);
-      for (const d of demos) {
-        expect(Object.isFrozen(d)).toBe(true);
-      }
-    }
-  });
-
-  it("if ALL mapped candidates fail with unreadable errors, push a CRITICAL warning to the sink", () => {
-    // The slug has multiple candidates — inject statSync failures for
-    // every candidate so the function exhausts all options.
-    const slug = "langgraph-typescript"; // mapped
-    const mapped = SLUG_TO_EXAMPLES[slug];
-    expect(mapped).toBeDefined();
-    // Create dirs for each candidate so existsSync succeeds (and we
-    // reach statSync where we inject).
-    for (const c of mapped!) makeExampleDir(root, c);
-    const targets = new Set(
-      mapped!.map((c) => path.join(root, "examples", "integrations", c)),
-    );
-    const orig = fs.statSync;
-    const spy = vi.spyOn(fs, "statSync").mockImplementation(((
-      p: fs.PathLike,
-      options?: unknown,
-    ) => {
-      if (typeof p === "string" && targets.has(p)) {
-        const e: NodeJS.ErrnoException = new Error("EACCES: injected");
-        e.code = "EACCES";
-        throw e;
-      }
-      return (orig as unknown as (p: fs.PathLike, o?: unknown) => unknown)(
-        p,
-        options,
-      );
-    }) as typeof fs.statSync);
-    try {
-      const cfg = makeConfig(root);
-      const sink: string[] = [];
-      const r = findExamplesSource(slug, cfg, sink);
-      expect(r.source).toBeNull();
-      expect(r.unreadableForSlug).toBe(true);
-      // A critical "all candidates unreadable" message must appear.
-      expect(sink.some((w) => /ERROR/.test(w))).toBe(true);
-      expect(sink.some((w) => w.includes(slug))).toBe(true);
-      expect(sink.some((w) => /unreadable/.test(w))).toBe(true);
-    } finally {
-      spy.mockRestore();
-    }
-  });
-});
-
-describe("buildReport — scalar summary fields", () => {
-  let root: string;
-  beforeEach(() => {
-    root = makeTmpTree();
-  });
-  afterEach(() => {
-    fs.rmSync(root, { recursive: true, force: true });
-  });
-
-  it("hasWarnings reflects whether any package has warnings", () => {
-    // A package whose mapped dir is missing emits a stale-mapping warning.
-    const mappedSlug = Object.keys(SLUG_TO_EXAMPLES)[0];
-    writePackage(root, mappedSlug, {
-      manifest: `slug: ${mappedSlug}\ndeployed: true\ndemos:\n  - id: a\n`,
-      specs: ["a.spec.ts"],
-      qaFiles: ["a.md"],
-    });
-    const cfg = makeConfig(root);
-    const report = buildReport([mappedSlug], cfg);
-    expect(report.hasWarnings).toBe(true);
-
-    // And a clean package — no warnings.
-    writePackage(root, "clean", {
-      manifest: `slug: clean\ndeployed: true\ndemos:\n  - id: a\n`,
-      specs: ["a.spec.ts"],
-      qaFiles: ["a.md"],
-    });
-    makeExampleDir(root, "clean");
-    const clean = buildReport(["clean"], cfg);
-    expect(clean.hasWarnings).toBe(false);
-  });
-});
-
-describe("parseArgs — --strict and --columns", () => {
-  it("parses --strict flag", () => {
-    const r = parseArgs(["--strict"]);
-    expect(r.strict).toBe(true);
-    expect(r.errors).toEqual([]);
-  });
-
-  it("--strict defaults to false", () => {
-    const r = parseArgs([]);
-    expect(r.strict).toBe(false);
-  });
-
-  it("parses --columns=a,b,c into an array of column keys", () => {
-    const r = parseArgs(["--columns=slug,demos,deployed"]);
-    expect(r.columns).toEqual(["slug", "demos", "deployed"]);
-    expect(r.errors).toEqual([]);
-  });
-
-  it("rejects unknown column keys in --columns", () => {
-    const r = parseArgs(["--columns=slug,bogus"]);
-    expect(r.errors.some((e) => /bogus/.test(e))).toBe(true);
-  });
-});
-
-describe("computeExitCode — --strict semantics", () => {
-  it("anomalies present → exit 1 regardless of strict/warnings", () => {
-    expect(
-      computeExitCode({
-        hasAnomalies: true,
-        hasWarnings: false,
-        strict: false,
-      }),
-    ).toBe(1);
-    expect(
-      computeExitCode({ hasAnomalies: true, hasWarnings: true, strict: false }),
-    ).toBe(1);
-    expect(
-      computeExitCode({ hasAnomalies: true, hasWarnings: true, strict: true }),
-    ).toBe(1);
-  });
-
-  it("no anomalies, warnings, default → exit 0 (default preserves current behavior)", () => {
-    expect(
-      computeExitCode({
-        hasAnomalies: false,
-        hasWarnings: true,
-        strict: false,
-      }),
-    ).toBe(0);
-  });
-
-  it("no anomalies, warnings, --strict → exit 5", () => {
-    expect(
-      computeExitCode({ hasAnomalies: false, hasWarnings: true, strict: true }),
-    ).toBe(5);
-  });
-
-  it("no anomalies, no warnings, --strict → exit 0 (strict only elevates when warnings exist)", () => {
-    expect(
-      computeExitCode({
-        hasAnomalies: false,
-        hasWarnings: false,
-        strict: true,
-      }),
-    ).toBe(0);
-  });
-
-  it("no anomalies, no warnings, default → exit 0", () => {
-    expect(
-      computeExitCode({
-        hasAnomalies: false,
-        hasWarnings: false,
-        strict: false,
-      }),
-    ).toBe(0);
-  });
-});
-
-describe("buildReport — --strict exit code", () => {
-  let root: string;
-
-  beforeEach(() => {
-    root = makeTmpTree();
-  });
-
-  afterEach(() => {
-    fs.rmSync(root, { recursive: true, force: true });
-  });
-
-  it("clean report with no warnings exits 0 regardless of --strict", () => {
-    // A fully clean package: manifest ok, deployed:true, counts match,
-    // examples dir present, and the slug is born-in-showcase so
-    // findExamplesSource is not consulted → no warnings, no anomalies.
-    const slug = "ag2"; // born-in-showcase → no SLUG_TO_EXAMPLES lookup
-    writePackage(root, slug, {
-      manifest: `slug: ${slug}\ndeployed: true\ndemos:\n  - id: a\n`,
-      specs: ["a.spec.ts"],
-      qaFiles: ["a.md"],
-    });
-    const cfg = makeConfig(root);
-
-    const relaxed = buildReport([slug], cfg);
-    expect(relaxed.hasWarnings).toBe(false);
-    expect(relaxed.hasAnomalies).toBe(false);
-    expect(relaxed.exitCode).toBe(0);
-
-    const strict = buildReport([slug], cfg, { strict: true });
-    expect(strict.hasWarnings).toBe(false);
-    expect(strict.hasAnomalies).toBe(false);
-    expect(strict.exitCode).toBe(0);
-  });
-
-  it("--strict with warnings-only (no anomalies) exits 5 (EXIT_WARNINGS)", () => {
-    // Construct the warnings-only quadrant deterministically by
-    // driving findExamplesSource into the "all-candidates-unreadable"
-    // path: mapped slug whose candidates exist on disk but whose
-    // statSync throws for each — findExamplesSource returns null AND
-    // pushes a CRITICAL warning onto the sink. auditPackage would
-    // also push a missing-examples anomaly unless the slug is in
-    // BORN_IN_SHOWCASE. To avoid that anomaly while keeping the
-    // warning, we mock SLUG_TO_EXAMPLES-driven behavior on a
-    // born-in-showcase slug — impossible since BORN_IN_SHOWCASE is
-    // the set of slugs without mappings.
-    //
-    // Instead, drive the warnings-only path by mocking statSync so
-    // findExamplesSource SUCCEEDS (dir statSync returns isDirectory)
-    // AND a warning is pushed to the sink. We achieve that by
-    // creating both candidates and forcing a stat failure on a LATER
-    // candidate while the FIRST resolves successfully:
-    //   - candidates[0] exists → returns path (no anomaly)
-    //   - but we first make the loop visit a pre-existing failing
-    //     candidate by creating it under the wrong name? No —
-    //     findExamplesSource iterates declared order and returns on
-    //     first match.
-    //
-    // Simpler deterministic route: mock fs.statSync once so the first
-    // candidate's stat throws EIO (pushing a statSync warning) AND the
-    // statSync-catch `continue` then finds no later candidate → returns
-    // null → missing-examples anomaly appears for non-BORN slugs.
-    //
-    // The cleanest deterministic warnings-only setup uses TWO packages:
-    // one born-in-showcase clean package plus a second package we mock
-    // to push a warning through findExamplesSource's sink without
-    // producing an anomaly. Since the anomaly-vs-warning coupling
-    // makes this hard to produce organically, we verify the --strict
-    // exit-code path directly via computeExitCode AND confirm that
-    // when BOTH warnings and anomalies coexist, --strict does NOT
-    // downgrade the anomaly exit code.
-    const slug = "mastra"; // mapped, missing dir → warning + anomaly
-    writePackage(root, slug, {
-      manifest: `slug: ${slug}\ndeployed: true\ndemos:\n  - id: a\n`,
-      specs: ["a.spec.ts"],
-      qaFiles: ["a.md"],
-    });
-    const cfg = makeConfig(root);
-    const report = buildReport([slug], cfg, { strict: true });
-    // Sanity: warnings AND anomalies both present in this path.
-    expect(report.hasWarnings).toBe(true);
-    expect(report.hasAnomalies).toBe(true);
-    // Anomaly exit wins over warnings exit (contract of computeExitCode).
-    expect(report.exitCode).toBe(1);
-
-    // Directly verify the warnings-only + strict exit path at the
-    // pure level (covered already in computeExitCode suite but
-    // asserted here for co-located clarity of the --strict contract).
-    expect(
-      computeExitCode({
-        hasAnomalies: false,
-        hasWarnings: true,
-        strict: true,
-      }),
-    ).toBe(5);
-    expect(
-      computeExitCode({
-        hasAnomalies: false,
-        hasWarnings: true,
-        strict: false,
-      }),
-    ).toBe(0);
-  });
-});
-
-describe("main() --columns via CLI subprocess", () => {
-  let root: string;
-
-  beforeEach(() => {
-    root = makeTmpTree();
-  });
-
-  afterEach(() => {
-    fs.rmSync(root, { recursive: true, force: true });
-  });
-
-  function runCli(
-    args: string[],
-    opts: { cwd?: string; env?: NodeJS.ProcessEnv } = {},
-  ) {
-    return spawnSync("npx", ["tsx", AUDIT_SCRIPT, ...args], {
-      cwd: opts.cwd,
-      env: { ...process.env, ...opts.env },
-      encoding: "utf-8",
-      timeout: 30_000,
-    });
-  }
-
-  it("--columns filters the table to the specified columns", () => {
-    writePackage(root, "crewai-crews", {
-      manifest: `slug: crewai-crews\ndeployed: true\ndemos:\n  - id: a\n`,
-      specs: ["a.spec.ts"],
-      qaFiles: ["a.md"],
-    });
-    makeExampleDir(root, "crewai-crews");
-    const r = runCli(["--columns=slug,demos"], {
-      env: { SHOWCASE_AUDIT_ROOT: root },
-    });
-    expect(r.status, r.stdout + r.stderr).toBe(0);
-    // Full columns include "deployed" and "examples src"; filtered
-    // output must NOT include those labels. Column headers checked
-    // as whole tokens — "slug" naked matches any line with the word
-    // (e.g. "crewai-crews" row) so pin a header-context regex.
-    expect(r.stdout).toMatch(/\bslug\b/);
-    expect(r.stdout).toMatch(/\bdemos\b/);
-    expect(r.stdout).not.toMatch(/\bdeployed\b/);
-    expect(r.stdout).not.toContain("examples src");
-    // The fixture slug must appear as its own data row.
-    expect(r.stdout).toMatch(/\bcrewai-crews\b/);
-    // And the Overall health summary must still render even with a
-    // filtered table.
-    expect(r.stdout).toMatch(/Packages total:\s+1/);
-  });
-});
-
-describe("module isMain guard", () => {
-  it("does not execute main() when imported as a subprocess (proof via spawnSync)", () => {
-    // Replace the tautological in-process assertion with a real
-    // subprocess test. We invoke node on a tiny inline script that
-    // imports audit.js (as a URL, since the real file is audit.ts and
-    // emits as audit.js in the module graph) and verifies it exits 0.
-    // If main() ran on import, it would exit 1 (empty packages) or 3
-    // (missing packages), not 0.
-    const helper = fs.mkdtempSync(path.join(os.tmpdir(), "audit-import-"));
-    const helperScript = path.join(helper, "probe.mjs");
-    // Use tsx to import the .ts file directly — tsx resolves the .js
-    // extension against the source .ts.
-    fs.writeFileSync(
-      helperScript,
-      `import("${AUDIT_SCRIPT.replace(/\\/g, "/")}").then((m) => {
-  if (typeof m.auditPackage !== "function") process.exit(1);
-  if (typeof m.buildReport !== "function") process.exit(1);
-  process.exit(0);
-}).catch((e) => {
-  console.error(e);
-  process.exit(2);
-});
-`,
-    );
-    try {
-      const r = spawnSync("npx", ["tsx", helperScript], {
-        encoding: "utf-8",
-        timeout: 30_000,
-      });
-      expect(r.status, r.stdout + r.stderr).toBe(0);
-    } finally {
-      fs.rmSync(helper, { recursive: true, force: true });
-    }
-  });
-});
